### 2024-09-16 3598065
chore: Update colorrange for coordinates in coordinate_example.jl
