+++ dev/assets/examples_coordinate_example.md.DjJEJNkG.js
@@ -1,54 +1,0 @@
-import{_ as i}from"./chunks/coordinate.Cu3U3WrH.js";import{_ as a,c as n,a2 as h,o as k}from"./chunks/framework.BGvApZjk.js";const F=JSON.parse('{"title":"Raytracing with inclination","description":"","frontmatter":{},"headers":[],"relativePath":"examples/coordinate_example.md","filePath":"examples/coordinate_example.md","lastUpdated":null}'),t={name:"examples/coordinate_example.md"};function l(p,s,e,E,r,d){return k(),n("div",null,s[0]||(s[0]=[h(`<h1 id="Raytracing-with-inclination" tabindex="-1">Raytracing with inclination <a class="header-anchor" href="#Raytracing-with-inclination" aria-label="Permalink to &quot;Raytracing with inclination {#Raytracing-with-inclination}&quot;">​</a></h1><p>In this example, we will raytrace the region around a Kerr blackhole as seen by an observer stationed at infinity. We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons as they are emitted from the source, at a fixed inclination angle from the blackhole&#39;s spin axis.</p><p>First, let&#39;s import Krang and CairoMakie for plotting.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
-<span class="line"></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">curr_theme </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Theme</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    Axis </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> (</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">        xticksvisible </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> false</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">        xticklabelsvisible </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> false</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
[…]
