### 2024-09-18 db0081c
chore: Update coordinates modulo operation in coordinate_example.jl
