### 2024-09-18 849d8f8
chore: Update coordinates modulo operation and colorrange in coordinate_example.jl
