### 2024-09-18 9d355b2
chore: Update emission coordinates modulo operation to use π instead of 2π
