+++ dev/assets/chunks/theme.CVFLS29I.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/VPLocalSearchBox.DXTZ5yk0.js","assets/chunks/framework.aDjGwPaf.js"])))=>i.map(i=>d[i]);
+import{d as m,o as a,c as u,r as c,n as I,a as O,t as w,b as g,w as f,e as h,T as de,_ as $,u as Ge,i as ze,f as Re,g as pe,h as y,j as v,k as r,l as R,m as re,p as T,q as F,s as Z,v as z,x as ve,y as fe,z as Ke,A as Je,B as K,F as M,C as B,D as Le,E as x,G as k,H as E,I as Te,J as ee,K as G,L as W,M as We,N as Ne,O as ie,P as we,Q as Ie,R as te,S as qe,U as Ye,V as Xe,W as Me,X as he,Y as Qe,Z as Ze,$ as xe,a0 as et,a1 as Ae,a2 as tt,a3 as nt,a4 as st,a5 as ye}from"./framework.aDjGwPaf.js";const ot=m({__name:"VPBadge",props:{text:{},type:{default:"tip"}},setup(o){return(e,t)=>(a(),u("span",{class:I(["VPBadge",e.type])},[c(e.$slots,"default",{},()=>[O(w(e.text),1)])],2))}}),at={key:0,class:"VPBackdrop"},rt=m({__name:"VPBackdrop",props:{show:{type:Boolean}},setup(o){return(e,t)=>(a(),g(de,{name:"fade"},{default:f(()=>[e.show?(a(),u("div",at)):h("",!0)]),_:1}))}}),it=$(rt,[["__scopeId","data-v-b06cdb19"]]),L=Ge;function lt(o,e){let t,s=!1;return()=>{t&&clearTimeout(t),s?t=setTimeout(o,e):(o(),(s=!0)&&setTimeout(()=>s=!1,e))}}function le(o){return/^\//.test(o)?o:`/${o}`}function me(o){const{pathname:e,search:t,hash:s,protocol:n}=new URL(o,"http://a.com");if(ze(o)||o.startsWith("#")||!n.startsWith("http")||!Re(e))return o;const{site:i}=L(),l=e.endsWith("/")||e.endsWith(".html")?o:o.replace(/(?:(^\.+)\/)?.*$/,`$1${e.replace(/(\.md)?$/,i.value.cleanUrls?"":".html")}${t}${s}`);return pe(l)}function Y({correspondingLink:o=!1}={}){const{site:e,localeIndex:t,page:s,theme:n,hash:i}=L(),l=y(()=>{var d,_;return{label:(d=e.value.locales[t.value])==null?void 0:d.label,link:((_=e.value.locales[t.value])==null?void 0:_.link)||(t.value==="root"?"/":`/${t.value}/`)}});return{localeLinks:y(()=>Object.entries(e.value.locales).flatMap(([d,_])=>l.value.label===_.label?[]:{text:_.label,link:ct(_.link||(d==="root"?"/":`/${d}/`),n.value.i18nRouting!==!1&&o,s.value.relativePath.slice(l.value.link.length-1),!e.value.cleanUrls)+i.value})),currentLang:l}}function ct(o,e,t,s){return e?o.replace(/\/$/,"")+le(t.replace(/(^|\/)index\.md$/,"$1").replace(/\.md$/,s?".html":"")):o}const ut={class:"NotFound"},dt={class:"code"},pt={class:"title"},vt={class:"quote"},ft={class:"action"},ht=["href","aria-label"],mt=m({__name:"NotFound",setup(o){const{theme:e}=L(),{
[…]

[…]
