### 2024-09-18 #7222 @dominic-chang
> **Submitting author:** @dominic-chang ([Dominic Chang](http://orcid.org/0000-0001-9939-5257)) **Repository:** https://github.com/dominic-chang/Krang.jl **Branch with paper.md** (empty if default branch): joss-paper **Version:** v0.1.0 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Dan Foreman-Mackey
>
> ## Status
> [![status](https://camo.githubusercontent.com/e1e388d4fca2cfa158d3b2ff597cef1565bb55f7fc241979e32ea7b168afe2ec/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f33373864663563353463643231653239336239326163363932633231633065642f7374617475732e737667)](https://joss.theoj.org/papers/378df5c54cd21e293b92ac692c21c0ed)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @dominic-chang. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @dominic-chang if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

I'd like to recommend alejandroc137 or aplavin as potential reviewers. I would like to point out that aplavin is at the same institution as me but did not work on this project.
