### 2024-09-18 177ef0a
Update coordinates modulo operation to use mod2pi function
