+++ dev/assets/chunks/@localSearchIndexroot.BEOY95H5.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":38,"nextId":38,"documentIds":{"0":"/Krang.jl/dev/api#Krang-api","1":"/Krang.jl/dev/api#Raytracing-Functions","2":"/Krang.jl/dev/api#Metric-Functions","3":"/Krang.jl/dev/api#Radial-Integrals","4":"/Krang.jl/dev/api#Angular-Integrals","5":"/Krang.jl/dev/api#Screen-Coordinates","6":"/Krang.jl/dev/api#Polarization-Functions","7":"/Krang.jl/dev/api#Raytracing-API-Related-Functions","8":"/Krang.jl/dev/api#misc","9":"/Krang.jl/dev/examples/coordinate_example#Raytracing-with-inclination","10":"/Krang.jl/dev/examples/mino_time_example#Rays-parameterized-with-Mino-time","11":"/Krang.jl/dev/examples/neural_net_example#Neural-Network-Emission-Model-Example","12":"/Krang.jl/dev/examples/neural_net_example#Fitting-the-model","13":"/Krang.jl/dev/examples/polarization_example#Creating-a-Custom-Dual-Cone-Model","14":"/Krang.jl/dev/getting_started#What-is-Krang.jl?","15":"/Krang.jl/dev/getting_started#philosophy","16":"/Krang.jl/dev/getting_started#raytracing","17":"/Krang.jl/dev/getting_started#Raytracing-conical-surfaces","18":"/Krang.jl/dev/getting_started#Raytracing-with-rays-parameterized-by-Mino-time","19":"/Krang.jl/dev/examples/raytracting_mesh_example#Raytracing-a-polygon-mesh","20":"/Krang.jl/dev/kerr_geodesic_summary#The-Geodesics-of-the-Kerr-Metric","21":"/Krang.jl/dev/kerr_geodesic_summary#Geodesic-Equations","22":"/Krang.jl/dev/kerr_geodesic_summary#Hamilton-Jacobi-Equations","23":"/Krang.jl/dev/kerr_geodesic_summary#Solutions-by-Quadrature","24":"/Krang.jl/dev/kerr_geodesic_summary#references","25":"/Krang.jl/dev/polarization#linear-polarization","26":"/Krang.jl/dev/polarization#Linear-Polarization-of-null-geodesics-in-the-Newmann-Penrose-Formalism","27":"/Krang.jl/dev/polarization#Walker-Penrose-Constant-and-Killing-Spinors-of-type-D-spacetimes","28":"/Krang.jl/dev/polarization#references","29":"/Krang.jl/dev/time_regularization#Regularized-time","30":"/Krang.jl/dev/time_regularization#references","31":"/Krang.jl/dev/newmann_penrose#Newmann-Penrose-Formalism","32":"/Krang.jl/dev/newmann_penrose#notation","33":"/Krang.jl/dev/newmann_penrose#Tetrad-Calculus","34":"/Krang.jl/dev/newmann_penrose#Spinoral-Calculus","35":"/Krang.jl/dev/newmann_penrose#Type-D-SpaceTimes","3
[…]
