### 2024-09-18 4ebe605
Fix plot range in examples
