+++ dev/assets/getting_started.md.C5sxIMbi.lean.js
@@ -1,0 +1,1 @@
+import{_ as o}from"./chunks/coordinate.C_naXwub.js";import{_ as n}from"./chunks/raytrace.D-ajFCJV.js";import{_ as l,c as i,a2 as s,j as t,a,o as r}from"./chunks/framework.aDjGwPaf.js";const M=JSON.parse('{"title":"What is Krang.jl?","description":"","frontmatter":{},"headers":[],"relativePath":"getting_started.md","filePath":"getting_started.md","lastUpdated":null}'),d={name:"getting_started.md"},m={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},p={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.355ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.999ex",height:"1.95ex",role:"img",focusable:"false",viewBox:"0 -705 883.6 862.1","aria-hidden":"true"},h={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.029ex"},xmlns:"http://www.w3.org/2000/svg",width:"3.054ex",height:"1.649ex",role:"img",focusable:"false",viewBox:"0 -716 1350 729","aria-hidden":"true"},g={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},u={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.023ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.061ex",height:"1.618ex",role:"img",focusable:"false",viewBox:"0 -705 469 715","aria-hidden":"true"};function x(T,e,Q,w,f,y){return r(),i("div",null,[e[11]||(e[11]=s('<h1 id="What-is-Krang.jl?" tabindex="-1">What is Krang.jl? <a class="header-anchor" href="#What-is-Krang.jl?" aria-label="Permalink to &quot;What is Krang.jl? {#What-is-Krang.jl?}&quot;">​</a></h1><p>Kerr Raytracer for Analytic Null Geodesics (Krang) is a raytracing code for geometries that are embedded in the Kerr spacetime. Krang solves the Kerr geodesic problem using the analytic solutions that have been derived from the Hamilton-Jacobi Formalism. This choice makes the code efficient and accurate, and is an ideal formalism to isolated sub images that manifest from strong lensing.</p><h2 id="philosophy" tabindex="-1">Philosophy <a class="header-anchor" href="#philosophy" aria-label="Permalink to &quot;Philosophy&quot;">​</a></h2><p>The current philosophy of Krang is to raytrace meshes. Meshes are made by &#39;painting&#39; materials onto geometries. These are then viewed by cameras that are sensitive to particular observables (<em>e.g. intensity and polarization</em>).</p><h2 id="raytracing" tabindex="-1">Raytracing <a class="header-anchor" href="#raytracing" aria-label="Permalink to &quot;Raytracing&quot;">​</a></h2>',5)),t("p",null,[e[4]||(e[4]=a("Light rays in this module can be parameterized in terms of either the cones (")),t("mjx-container",m,[(r(),i("svg",p,e[0]||(e[0]=[s('<g stroke="currentColor" fill="currentColor" stroke-width="0" transform="scale(1,-1)"><g data-mml-node="math"><g data-mml-node="ms
[…]
