### 2024-09-18 c81da7c
Update camera and plot range in mino_time_example.jl
