### 2024-09-22 #7222 @lrnv
@alejandroc137, @aplavin, would you guys agree to review this submission for JOSS ? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
