### 2024-09-23 #7222 @lrnv
@ExpandingMan, would you be interested in doing a review of this submission for JOSS ? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
