### 2024-09-23 #7222 @alejandroc137
Hi! Yes, I will be happy to review this submission. Please let me know what I should do next. On Sep 22, 2024, at 5:46 AM, Oskar Laverny ***@***.***> wrote:﻿
@alejandroc137, @aplavin, would you guys agree to review this submission for JOSS ? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
