### 2024-09-24 e686a17
a couple of tiny performance optimizations
