### 2024-09-24 #7273 @editorialbot
[REVIEW]: Krang: Kerr Raytracer for Analytic Null Geodesics
**Submitting author:** @dominic-chang (<a href="http://orcid.org/0000-0001-9939-5257">Dominic Chang</a>)
**Repository:** https://github.com/dominic-chang/Krang.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.1
