### 2024-09-24 #7273 @lrnv
👋🏼 @dominic-chang, @aplavin, @ExpandingMan, @alejandroc137: this is the review thread for the paper. All of our communications will happen here from now on, so that each of us can read what each other writes.

@aplavin, @ExpandingMan, @alejandroc137, as reviewers, your first step is to create a checklist for yourself, to guide you through the review, by commenting

on this thread.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. **As such, the reviewers are encouraged to submit issues and pull requests on the software repository**. When doing so, **please mention openjournals/joss-reviews#7273** so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use @editorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me (@lrnv) if you have any questions/concerns, if something is unclear, or if you need more info about the process :)
