+++ src/Krang.jl
@@ -22,7 +22,6 @@ include("metrics/Kerr/misc.jl")
 include("metrics/Kerr/emission_coordinates.jl")
 include("metrics/Kerr/api.jl")
 include("materials/physicsUtils.jl")
-include("materials/material_types.jl")
 include("materials/CoordinateRadius.jl")
 include("materials/CoordinatePoint.jl")
 include("materials/ElectronSynchrotronPowerLawPolarization.jl")
