+++ src/materials/material_types.jl
@@ -1,6 +1,0 @@
-"""
-    $TYPEDEF
-
-Abstract Material
-"""
-abstract type AbstractMaterial end
