+++ examples/neural_net_example.jl
@@ -163,7 +163,7 @@ ps_trained, st_trained = let st=Ref(st), x=pixels, y=target_img
     
     solution = Optimization.solve(
         optprob,
-        OptimizationOptimisers.ADAM(0.05),
+        OptimizationOptimisers.Adam(0.05),
         maxiters = 1_000, 
     callback=Callback(100,()->nothing)
     )
