### 2024-09-26 dominic-chang/Krang.jl#20 @ExpandingMan
docs/examples suggestions
I've been taking a look at this package as part of [#7273](https://github.com/openjournals/joss-reviews/issues/7273) and have a few suggestions for things that I feel could benefit the documentation and examples.

For the most part, I think the source code is very clear (especially when compared to the referenced paper that gives the geodesic solutions) and not in need of a lot of extra exposition, but here are a few minor things I felt could use some improvement.

### Type System
I think it would be nice to add a brief section, for example [somewhere in here](https://dominic-chang.com/Krang.jl/dev/getting_started) describing how the various types defined in the package are used.  For example, it at first looks a little bit strange that in some cases raytracing is done by using the `CoordinatePoint` material, which apparently indicates that it should return a coordinate instead of intensity (it at first looks a bit strange that `CoordinatePoint` is an `AbstractMaterial`).  Perhaps also a brief mention of the various cameras, specifically what they can be used to compute, and that they are always located on the surface at infinity would be useful.

### Mesh Embeddings
It may be partially a result of my own ignorance concerning typical astrophysics applications, but at least from a general theoretical GR standpoint it is not at all obvious how you propose to embed imported meshes into the Kerr spacetime.  From looking at the code, it seems to me that you do this by identifying spherical polar coordinates in Euclidean space with the corresponding Boyer-Lindquist coordinates which (again, from an abstract geometric point of view) is not a particularly obvious thing to do (though I suppose it only starts to get real weird inside the horizon).  I think it would be good to explain this in some detail in the docs.

### Example Docs
In some cases I think it would be nice to be a little more verbose about what exactly is happening in the examples, particularly when there are large blocks of code uninterrupted by an explanation.  Some of this relates to the things I just mentioned above, in others I think it would be good to link to the theory sections of your own docs or to the original references (for example when `ElectronSynchrotronPowerLawPolarization` is introduced it might be nice to reference that paper right in the example so anyone who's interested can immediately find out exactly what that is).  Another example is in the neural network example where you are fitting to a random field generated by the randomly initialized network: in retrospect it's obvious what's being done, but it looked strange at first glance, since it'd only really make sense to do that for pedagogical reasons.

### GPU Example
I think if you are going to mention GPU's right in the [landing page](https://dominic-chang.com/Krang.jl/dev/) it would be a really good idea to show another short example that uses a GPU explicitly.  From what I can tell so far, it does indeed seem like the vast majority of this package should have no trouble running in GPU kernels, but we all know that GPU's are a gigantic pain in the ass and there will inevitably be a lot of "gotchas" and "you're holding it wrong" moments trying to get something to run with a GPU.  This is doubly true for anyone who is not already an experienced Julia user (and an experienced Julia GPU user, at that).  I realize that an example will only be able to cover one or two very narrow use cases, but, again, I feel there should be one if GPU is advertised.

----
Anyway, thanks for your work on this package!
