### 2024-09-27 dominic-chang/Krang.jl#20 @dominic-chang
Thanks so much. These are actually great suggestions!

> **Type System**
I think it would be nice to add a brief section, for example [somewhere in here](https://dominic-chang.com/Krang.jl/dev/getting_started) describing how the various types defined in the package are used. For example, it at first looks a little bit strange that in some cases raytracing is done by using the CoordinatePoint material, which apparently indicates that it should return a coordinate instead of intensity (it at first looks a bit strange that CoordinatePoint is an AbstractMaterial). Perhaps also a brief mention of the various cameras, specifically what they can be used to compute, and that they are always located on the surface at infinity would be useful.

Yes, I think this is a good point, especially since some of those types have different optimizations about them even if they can be interchanged. The CoordinatePoint material is admittedly kind of strange. I was trying to figure out how to plot coordinates while keeping to the philosophy of the code to, 'raytrace meshes made from materials that are painted onto geometries'. Perhaps I don't have to be that strict about this and I could emphasize to the reader that these are internal functions to return coordinate locations.

> **Mesh Embeddings**
It may be partially a result of my own ignorance concerning typical astrophysics applications, but at least from a general theoretical GR standpoint it is not at all obvious how you propose to embed imported meshes into the Kerr spacetime. From looking at the code, it seems to me that you do this by identifying spherical polar coordinates in Euclidean space with the corresponding Boyer-Lindquist coordinates which (again, from an abstract geometric point of view) is not a particularly obvious thing to do (though I suppose it only starts to get real weird inside the horizon). I think it would be good to explain this in some detail in the docs.

Yes, I would agree that the method for embedding that I have chosen is probably not the most natural choice since it's coordinate dependent. I think the most natural choice would to probably embed the mesh such that it's Gaussian curvature is preserved. This feature was added because it was something I was often using for benchmarking/testing, and thought that  it might be useful for someone else in the future. Maybe I can add some text explaining what's done and the caveats about it.

> **Example Docs**
In some cases I think it would be nice to be a little more verbose about what exactly is happening in the examples, particularly when there are large blocks of code uninterrupted by an explanation. Some of this relates to the things I just mentioned above, in others I think it would be good to link to the theory sections of your own docs or to the original references (for example when ElectronSynchrotronPowerLawPolarization is introduced it might be nice to reference that paper right in the example so anyone who's interested can immediately find out exactly what that is). Another example is in the neural network example where you are fitting to a random field generated by the randomly initialized network: in retrospect it's obvious what's being done, but it looked strange at first glance, since it'd only really make sense to do that for pedagogical reasons.

Yes, this is a great suggestion. I did want to add a more realistic neural network fitting example to the docs, but it would have required extra dependencies for handling actual data. I chose instead to keep it pedagogical. I'll highlight this.

> **GPU Example**
I think if you are going to mention GPU's right in the [landing page](https://dominic-chang.com/Krang.jl/dev/) it would be a really good idea to show another short example that uses a GPU explicitly.

I think this is a good idea but am not sure how to handle it. I never added any GPU examples because I didn't have access to a GPU machine with github CI. I think I can likely still add an example, but the docs will not be generated dynamically. Maybe that's the best approach for now?
