### 2024-09-27 dominic-chang/Krang.jl#20 @ExpandingMan
For the mesh embeddings, like I said, I'm more a theory person and my practical astrophysics knowledge is sketchy, so I was wondering if perhaps this choice is motivated by it?  I guess it's probably a convenient way to talk about accretion disks and jets and stuff, so I figure it's more practically useful than some fancy embedding that preserves some abstract quantity.  Any embedding you can choose is going to be arbitrary.  Again, nothing wrong with it, just in need of an explanation

As for GPU, I think it's fine if you don't test it in CI/CD if it's in example or docs, that's typically what people do.  If you can come up with an example I can help you verify it (at least for CUDA).  In particular I think it would be great to highlight if there are any functions here that can immediately benefit from being run on a `GPUArray`, and if not, what can people do.  I'm confident most of the stuff in this package will work fine in principle, but GPU's are so fiddly and annoying that even with a fair amount of experience of GPU's in Julia I'm not even a priori sure what the right approach for this package would be.
