### 2024-09-27 c702845
Update docs to include some description of various types
