### 2024-09-29 7db3afd
Update polarization example
