+++ dev/assets/examples_polarization_example.md.Hl9nzwdK.lean.js
@@ -1,0 +1,45 @@
+import{_ as e,c as n,a2 as a,j as i,a as t,o as l}from"./chunks/framework.DJ9CQGGt.js";const h="/Krang.jl/dev/assets/polarization_example.DRqwADdU.png",Q=JSON.parse('{"title":"Creating a Custom Dual Cone Model","description":"","frontmatter":{},"headers":[],"relativePath":"examples/polarization_example.md","filePath":"examples/polarization_example.md","lastUpdated":null}'),p={name:"examples/polarization_example.md"},k={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.023ex",height:"1.581ex",role:"img",focusable:"false",viewBox:"0 -677 1778 699","aria-hidden":"true"},d={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},E={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.186ex"},xmlns:"http://www.w3.org/2000/svg",width:"8.426ex",height:"1.808ex",role:"img",focusable:"false",viewBox:"0 -717 3724.1 799","aria-hidden":"true"},o={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"3.25ex",height:"1.672ex",role:"img",focusable:"false",viewBox:"0 -717 1436.6 739","aria-hidden":"true"};function y(c,s,F,C,m,u){return l(),n("div",null,[s[13]||(s[13]=a(`<h1 id="Creating-a-Custom-Dual-Cone-Model" tabindex="-1">Creating a Custom Dual Cone Model <a class="header-anchor" href="#Creating-a-Custom-Dual-Cone-Model" aria-label="Permalink to &quot;Creating a Custom Dual Cone Model {#Creating-a-Custom-Dual-Cone-Model}&quot;">​</a></h1><p>We will defined a custom model for low luminosity active galactice nuclei (LLAGN). A detailed description of the model can be found in this <a href="https://doi.org/10.48550/arXiv.2405.04749" target="_blank" rel="noreferrer">reference</a>. We will show the emission of the n=0 (direct) and n=1 (indirect) photons as they are emitted from the source, at a fixed inclination angle from the blackhole&#39;s spin axis.</p><p>First, let&#39;s import Krang and CairoMakie for plotting.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
+<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
+<span class="line"></span>
[…]
