+++ dev/assets/chunks/@localSearchIndexroot.Bctuss0A.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":42,"nextId":42,"documentIds":{"0":"/Krang.jl/dev/api#Krang-api","1":"/Krang.jl/dev/api#Raytracing-Functions","2":"/Krang.jl/dev/api#Metric-Functions","3":"/Krang.jl/dev/api#Radial-Integrals","4":"/Krang.jl/dev/api#Angular-Integrals","5":"/Krang.jl/dev/api#Screen-Coordinates","6":"/Krang.jl/dev/api#Polarization-Functions","7":"/Krang.jl/dev/api#Raytracing-API-Related-Functions","8":"/Krang.jl/dev/api#misc","9":"/Krang.jl/dev/examples/coordinate_example#Raytracing-with-inclination","10":"/Krang.jl/dev/examples/mino_time_example#Rays-parameterized-with-Mino-time","11":"/Krang.jl/dev/examples/polarization_example#Creating-a-Custom-Dual-Cone-Model","12":"/Krang.jl/dev/examples/neural_net_example#Neural-Network-Emission-Model-Example","13":"/Krang.jl/dev/examples/neural_net_example#Fitting-the-model","14":"/Krang.jl/dev/examples/raytracting_mesh_example#Raytracing-a-polygon-mesh","15":"/Krang.jl/dev/getting_started#What-is-Krang.jl?","16":"/Krang.jl/dev/getting_started#philosophy","17":"/Krang.jl/dev/getting_started#meshes","18":"/Krang.jl/dev/getting_started#materials","19":"/Krang.jl/dev/getting_started#geometries","20":"/Krang.jl/dev/getting_started#raytracing","21":"/Krang.jl/dev/getting_started#Raytracing-conical-surfaces","22":"/Krang.jl/dev/getting_started#Raytracing-with-rays-parameterized-by-Mino-time","23":"/Krang.jl/dev/getting_started#cameras","24":"/Krang.jl/dev/kerr_geodesic_summary#The-Geodesics-of-the-Kerr-Metric","25":"/Krang.jl/dev/kerr_geodesic_summary#Geodesic-Equations","26":"/Krang.jl/dev/kerr_geodesic_summary#Hamilton-Jacobi-Equations","27":"/Krang.jl/dev/kerr_geodesic_summary#Solutions-by-Quadrature","28":"/Krang.jl/dev/kerr_geodesic_summary#references","29":"/Krang.jl/dev/newmann_penrose#Newmann-Penrose-Formalism","30":"/Krang.jl/dev/newmann_penrose#notation","31":"/Krang.jl/dev/newmann_penrose#Tetrad-Calculus","32":"/Krang.jl/dev/newmann_penrose#Spinoral-Calculus","33":"/Krang.jl/dev/newmann_penrose#Type-D-SpaceTimes","34":"/Krang.jl/dev/newmann_penrose#Ricci-Rotation-Coefficients","35":"/Krang.jl/dev/newmann_penrose#references","36":"/Krang.jl/dev/polarization#linear-polarization","37":"/Krang.jl/dev/polarization#Linear-Polarization-of-null-geodesics-in-t
[…]

[…]
