### 2024-09-29 e3c787e
Add threaded render! method
