### 2024-09-29 135993e
refactor: Update neural network architecture and optimization parameters
