### 2024-09-29 cc152ff
Replace loss function in neural network example with mse
