### 2024-09-29 dbc15de
Update neural network example text
