### 2024-09-30 dominic-chang/Krang.jl#21 @aplavin
UndefVarError: `generate_ray` not defined
Trying out Krang as part of https://github.com/openjournals/joss-reviews/issues/7273, and noticed that the slow light example (https://dominic-chang.com/Krang.jl/dev/examples/mino_time_example) doesn't work: throws `UndefVarError: generate_ray not defined` as `Krang.generate_ray` doesn't seem to exist.
