### 2024-09-30 f4a014b
Add KernelAbstractions for render! call to use GPU. Updated coordinate example to mention GPU usage
