+++ src/schemes/schemes.jl
@@ -1,4 +1,4 @@
-export render, render!
+export render, render!, render_cpu_threaded!
 abstract type AbstractScheme end
 
 function render(camera::AbstractCamera, scene::Scene)  
@@ -10,7 +10,25 @@ function render(camera::AbstractCamera, scene::Scene)
     )
 end
 
+@kernel function _render!(store, pixels, mesh::Mesh)
+    I = @index(Global)
+    for I in CartesianIndices(pixels)
+        @inbounds store[I] = mesh.material(pixels[I], mesh.geometry)
+    end
+end
+
 function render!(store, camera::AbstractCamera, scene::Scene)  
+    backend = get_backend(store)
+    @assert backend == get_backend(camera.screen.pixels)
+    @assert size(store) == size(camera.screen.pixels)
+    mapreduce(
+        mesh -> begin _render!(backend)(store, camera.screen.pixels, mesh, ndrange=size(store)); store end,
+        +,
+        scene
+    )
+end
+
+function render_cpu_threaded!(store, camera::AbstractCamera, scene::Scene)  
     @assert size(store) == size(camera.screen.pixels)
     mapreduce(
         mesh -> begin
