### 2024-10-01 f006525
Fix Looping bug when rendering
