+++ src/cameras/SlowLightIntensityCamera.jl
@@ -73,8 +73,8 @@ struct SlowLightIntensityScreen{T, A <:AbstractMatrix} <: AbstractScreen
 
     @kernel function _generate_screen!(screen, met::Kerr{T}, αmin, αmax, βmin, βmax, θo, res) where T
         I,J = @index(Global, NTuple)
-        α = αmin + (αmax - αmin) * T(I) / res
-        β = βmin + (βmax - βmin) * T(J) / res
+        α = αmin + (αmax - αmin) * (T(I)-1) / (res-1)
+        β = βmin + (βmax - βmin) * (T(J)-1) / (res-1)
         screen[I, J] = SlowLightIntensityPixel(met, α, β, θo)
     end
     function SlowLightIntensityScreen(met::Kerr{T}, αmin, αmax, βmin, βmax, θo, res; A=Matrix) where {T}
