### 2024-10-01 1e52ce2
Update Project.toml with Enzyme compatibility
