+++ Project.toml
@@ -20,7 +20,7 @@ PolarizedTypes = "0.1"
 Rotations = "1"
 StaticArrays = "1"
 julia = "1.10"
-KernelAbstractions = "0.9.27"
+KernelAbstractions = "0.9"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
