### 2024-10-01 dominic-chang/Krang.jl#20 @dominic-chang
I think I should have address all of these in the most recent version of the docs
