+++ v0.2.0/assets/chunks/coordinate.JMLhO2E7.js
@@ -1,0 +1,1 @@
+const s="/Krang.jl/v0.2.0/assets/coordinate.BHWT5GLC.gif";export{s as _};
