### 2024-10-01 00daacc
fixes for initializing GPU pixel array
