### 2024-10-04 c2090e8
Refactor coordinate_example.jl and remove CoordinatePoint and CoordinateRadius materials
