### 2024-10-04 8c6ae4b
Bump version to 0.2.1 in Project.toml Refactor coordinate_example.jl: Remove CoordinatePoint material and update camera initialization
