+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Krang"
 uuid = "54806c32-d51a-438d-8447-e0041be2fbfb"
 authors = ["Dominic <dchang3419@hotmail.com> and contributors"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
