+++ dev/assets/examples_raytracting_mesh_example.md.jxQoIF07.lean.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,a2 as n,o as h}from"./chunks/framework.DJ9CQGGt.js";const k="/Krang.jl/dev/assets/stanford_bunny.DaYXOfLO.png",t="/Krang.jl/dev/assets/mesh_geometry_example.CFinjHq3.png",F=JSON.parse('{"title":"Raytracing a polygon mesh","description":"","frontmatter":{},"headers":[],"relativePath":"examples/raytracting_mesh_example.md","filePath":"examples/raytracting_mesh_example.md","lastUpdated":null}'),p={name:"examples/raytracting_mesh_example.md"};function l(e,s,E,r,d,g){return h(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Raytracing-a-polygon-mesh" tabindex="-1">Raytracing a polygon mesh <a class="header-anchor" href="#Raytracing-a-polygon-mesh" aria-label="Permalink to &quot;Raytracing a polygon mesh {#Raytracing-a-polygon-mesh}&quot;">​</a></h1><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
+import{_ as i,c as a,a2 as n,o as h}from"./chunks/framework.Bt44EZ1S.js";const k="/Krang.jl/dev/assets/stanford_bunny.DaYXOfLO.png",t="/Krang.jl/dev/assets/mesh_geometry_example.CFinjHq3.png",F=JSON.parse('{"title":"Raytracing a polygon mesh","description":"","frontmatter":{},"headers":[],"relativePath":"examples/raytracting_mesh_example.md","filePath":"examples/raytracting_mesh_example.md","lastUpdated":null}'),p={name:"examples/raytracting_mesh_example.md"};function l(e,s,E,r,d,g){return h(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Raytracing-a-polygon-mesh" tabindex="-1">Raytracing a polygon mesh <a class="header-anchor" href="#Raytracing-a-polygon-mesh" aria-label="Permalink to &quot;Raytracing a polygon mesh {#Raytracing-a-polygon-mesh}&quot;">​</a></h1><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
 <span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">import</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">as</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CMk</span></span>
 <span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeometryBasics</span></span>
 <span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> FileIO</span></span>
