+++ v0.2.1/assets/chunks/@localSearchIndexroot.BGL-AZbH.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":42,"nextId":42,"documentIds":{"0":"/Krang.jl/v0.2.1/api#Krang-api","1":"/Krang.jl/v0.2.1/api#Raytracing-Functions","2":"/Krang.jl/v0.2.1/api#Metric-Functions","3":"/Krang.jl/v0.2.1/api#Radial-Integrals","4":"/Krang.jl/v0.2.1/api#Angular-Integrals","5":"/Krang.jl/v0.2.1/api#Screen-Coordinates","6":"/Krang.jl/v0.2.1/api#Polarization-Functions","7":"/Krang.jl/v0.2.1/api#Raytracing-API-Related-Functions","8":"/Krang.jl/v0.2.1/api#misc","9":"/Krang.jl/v0.2.1/examples/coordinate_example#Raytracing-with-inclination","10":"/Krang.jl/v0.2.1/examples/mino_time_example#Rays-parameterized-with-Mino-time","11":"/Krang.jl/v0.2.1/examples/neural_net_example#Neural-Network-Emission-Model-Example","12":"/Krang.jl/v0.2.1/examples/neural_net_example#Fitting-the-model","13":"/Krang.jl/v0.2.1/examples/polarization_example#Creating-a-Custom-Dual-Cone-Model","14":"/Krang.jl/v0.2.1/examples/raytracting_mesh_example#Raytracing-a-polygon-mesh","15":"/Krang.jl/v0.2.1/getting_started#What-is-Krang.jl?","16":"/Krang.jl/v0.2.1/getting_started#philosophy","17":"/Krang.jl/v0.2.1/getting_started#meshes","18":"/Krang.jl/v0.2.1/getting_started#materials","19":"/Krang.jl/v0.2.1/getting_started#geometries","20":"/Krang.jl/v0.2.1/getting_started#raytracing","21":"/Krang.jl/v0.2.1/getting_started#Raytracing-conical-surfaces","22":"/Krang.jl/v0.2.1/getting_started#Raytracing-with-rays-parameterized-by-Mino-time","23":"/Krang.jl/v0.2.1/getting_started#cameras","24":"/Krang.jl/v0.2.1/kerr_geodesic_summary#The-Geodesics-of-the-Kerr-Metric","25":"/Krang.jl/v0.2.1/kerr_geodesic_summary#Geodesic-Equations","26":"/Krang.jl/v0.2.1/kerr_geodesic_summary#Hamilton-Jacobi-Equations","27":"/Krang.jl/v0.2.1/kerr_geodesic_summary#Solutions-by-Quadrature","28":"/Krang.jl/v0.2.1/kerr_geodesic_summary#references","29":"/Krang.jl/v0.2.1/newmann_penrose#Newmann-Penrose-Formalism","30":"/Krang.jl/v0.2.1/newmann_penrose#notation","31":"/Krang.jl/v0.2.1/newmann_penrose#Tetrad-Calculus","32":"/Krang.jl/v0.2.1/newmann_penrose#Spinoral-Calculus","33":"/Krang.jl/v0.2.1/newmann_penrose#Type-D-SpaceTimes","34":"/Krang.jl/v0.2.1/newmann_penrose#Ricci-Rotation-Coefficients","35":"/Krang.jl/v0.2.1/newmann_penrose#references","36":"/Krang.jl/v0.2.1/polarizat
[…]
