### 2024-10-05 f25fd17
Bump version to 0.2.2 in Project.toml
