### 2024-10-07 dominic-chang/Krang.jl#20 @alejandroc137
I just started to go over the code. For now, just a couple of comments on the examples/documentation:

- Since the examples make heavy use of CairoMakie, I think it should be added as a dependence.
- I think it is worth adding some notebooks instead of only scripts, at least for the most basic examples. This will allow for more comments and embedded figures that give a sense of what is going on without even running the code.
- The documentation has several typos (e.g., blackhole -> black hole, assymptotic-> asymptotic, etc). It may be worth putting the text on a word processor to check it.

I run https://dominic-chang.com/Krang.jl/v0.2.0/examples/coordinate_example and received this error when running

`material = Krang.CoordinatePoint();`

It returns this:

```
UndefVarError: `CoordinatePoint` not defined

Stacktrace:
 [1] getproperty(x::Module, f::Symbol)
   @ Base ./Base.jl:31
 [2] top-level scope
   @ In[5]:2
```

Is this fixed in the latest version (I installed it from the Pkg mode)?
