### 2024-10-07 992af32
fix coordinate_example docs
