### 2024-10-07 ea842d6
Fix typo in index.md
