### 2024-10-07 afb17d4
Update link to coordinate example in navigation
