+++ dev/assets/examples_neural_net_example.md.DrCtVll1.lean.js
@@ -1,196 +1,0 @@
-import{_ as p,c as n,a2 as a,j as i,a as h,o as l}from"./chunks/framework.Bt44EZ1S.js";const k="/Krang.jl/dev/assets/emission_model_and_target_model.CHGfVBVO.png",t="/Krang.jl/dev/assets/emission_model_and_image_model.Q-7T4rZy.png",e="/Krang.jl/dev/assets/neural_net_results.D5hefsYy.png",A=JSON.parse('{"title":"Neural Network Emission Model Example","description":"","frontmatter":{},"headers":[],"relativePath":"examples/neural_net_example.md","filePath":"examples/neural_net_example.md","lastUpdated":null}'),E={name:"examples/neural_net_example.md"},r={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"9.517ex",height:"2.452ex",role:"img",focusable:"false",viewBox:"0 -833.9 4206.6 1083.9","aria-hidden":"true"};function g(y,s,F,o,C,c){return l(),n("div",null,[s[4]||(s[4]=a(`<h1 id="Neural-Network-Emission-Model-Example" tabindex="-1">Neural Network Emission Model Example <a class="header-anchor" href="#Neural-Network-Emission-Model-Example" aria-label="Permalink to &quot;Neural Network Emission Model Example {#Neural-Network-Emission-Model-Example}&quot;">​</a></h1><p>This is a pedagogical example that serves as a proof of concept. We will build a simple Neural Network emission model to be ray traced and optimize with ADAM algorithm.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Lux</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Random</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">Random</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">seed!</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">123</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span>
[…]
