### 2024-10-07 c599de4
Bump version to 0.2.3 in Project.toml
