+++ v0.2.3/assets/examples_mino-time-example.md.DPI2Zncc.js
@@ -1,0 +1,72 @@
+import{_ as i}from"./chunks/raytrace.CHI3Lgya.js";import{_ as a,c as h,a2 as k,o as n}from"./chunks/framework.I6k3UwvE.js";const t="/Krang.jl/v0.2.3/assets/mino_time_rays.BIotIXNT.png",C=JSON.parse('{"title":"Rays parameterized with Mino time","description":"","frontmatter":{},"headers":[],"relativePath":"examples/mino-time-example.md","filePath":"examples/mino-time-example.md","lastUpdated":null}'),l={name:"examples/mino-time-example.md"};function p(e,s,E,r,d,g){return n(),h("div",null,s[0]||(s[0]=[k(`<h1 id="Rays-parameterized-with-Mino-time" tabindex="-1">Rays parameterized with Mino time <a class="header-anchor" href="#Rays-parameterized-with-Mino-time" aria-label="Permalink to &quot;Rays parameterized with Mino time {#Rays-parameterized-with-Mino-time}&quot;">​</a></h1><p>In this example, we will raytrace the region around a Kerr black hole as seen by an observer stationed at infinity. We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons as they are emitted from the source, at a fixed inclination angle from the black hole&#39;s spin axis.</p><p>First, let&#39;s import Krang and CairoMakie for plotting.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
+<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">import</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GLMakie </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">as</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GLMk</span></span>
[…]
