### 2024-10-08 3c3e2e1
Fix missing gif in coordinate-examle
