### 2024-10-08 dominic-chang/Krang.jl#20 @dominic-chang
Thanks for these comments

> - Since the examples make heavy use of CairoMakie, I think it should be added as a dependence.

CairoMakie is quite a heavy package and is really only used in the examples for plotting. Adding it as a dependency would increase the package pre-compilation time by quite a bit, for not much benefit, especially since different users may use different plotting packages. I do think it's a good idea to add plotting recipes for `Makie` and `Plots`, but this will likely be done through a package extension, meaning that users would still need to install those packages separately.

> - I think it is worth adding some notebooks instead of only scripts, at least for the most basic examples. This will allow for more comments and embedded figures that give a sense of what is going on without even running the code.

Sure, this is a good idea. I can try and transcribe the examples into notebooks for usage. I think this also implies that I should add more text explaining what's going on in the examples on the page. Please feel free to point out anything that you think is particularly obtuse.

> - The documentation has several typos (e.g., blackhole -> black hole, assymptotic-> asymptotic, etc). It may be worth putting the text on a word processor to check it.

Thanks for pointing this out. I ran the site through a spell check and I believe I've gotten everything now.

> I run https://dominic-chang.com/Krang.jl/v0.2.0/examples/coordinate_example and received this error when running
> `material = Krang.CoordinatePoint();`

I removed this material in the last version. This should be fixed in the latest docs (https://dominic-chang.com/Krang.jl/stable/examples/coordinate-example). I would also run `up` in pkg mode to make sure your on the latest version.
