+++ dev/assets/examples_coordinate-example.md.cYk-3wLI.js
@@ -1,4 +1,4 @@
-import{_ as k,c as n,a2 as a,j as i,a as h,o as t}from"./chunks/framework.CCA2VXdB.js";const c=JSON.parse('{"title":"Raytracing with inclination","description":"","frontmatter":{},"headers":[],"relativePath":"examples/coordinate-example.md","filePath":"examples/coordinate-example.md","lastUpdated":null}'),l={name:"examples/coordinate-example.md"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},e={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.355ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.999ex",height:"1.95ex",role:"img",focusable:"false",viewBox:"0 -705 883.6 862.1","aria-hidden":"true"};function r(E,s,d,g,y,o){return t(),n("div",null,[s[4]||(s[4]=a(`<h1 id="Raytracing-with-inclination" tabindex="-1">Raytracing with inclination <a class="header-anchor" href="#Raytracing-with-inclination" aria-label="Permalink to &quot;Raytracing with inclination {#Raytracing-with-inclination}&quot;">​</a></h1><p>This example shows how to access coordinate information from the raytracing process. You will likely need to do this when making custom physics <code>materials</code>. We will raytrace a sequence of cones in the region around a Kerr black hole as seen by an observer stationed at infinity. We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons as they are emitted from the source, at a fixed inclination angle from the black hole&#39;s spin axis.</p><p>First, let&#39;s import Krang and CairoMakie for plotting.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
[…]
