### 2024-10-08 dominic-chang/Krang.jl#20 @ExpandingMan
Just want to chime in and say that it is definitely *not* standard practice in Julia to add plotting packages like CairoMakie as dependencies.  Indeed as Dominic says, they are quite "heavy", not to mention they can potentially add version caps in the environment.  It's very easy to install as a separate package and use in an example.
