### 2024-10-08 dominic-chang/Krang.jl#20 @alejandroc137
Thank you both (@dominic-chang and @ExpandingMan) for the clarification. I think it is still worth clearly mentioning the dependencies all the examples need (e.g., CairoMakie or GLMakie), as that is indeed a standard practice.

- I indeed think you may want to comment a little bit more on the examples. Going back to the coordinate example, e.g., small things like just commenting on these variables will be helpful (which you are doing only in the second example):

sze = 400; # Will always the screen be a square of size \sqrt(sze)?
rmin = Krang.horizon(metric)
rmax = 10.0;
ρmax = 15.0; # Is this something related to the observer's screen? How does this related to sze?

Since these appear in most cases and are very common, you may want to do it somewhere in a very general way. For example, in the "getting started," it can be nice to add the most general parameters and their meaning.

- Please also mention the bounds on these parameters that you know the code will work. For example, will it work for θo=0, θo=90, a=0.0 or a=1.0? It seems these parameters are allowed, and nothing seems to crash (you even use in one example θo = 90/180*π). However, I am not sure the analytic expressions work correctly on the bounds. It would also be nice to add other physical restrictions explicitly; for example, are there any limitations on the opening angle of the cones? Do the dual cones always have to have the same opening angles?

- I think it would be very useful if you also provided an explicit example of how to export the results of the raytracing and other results. Users may want to export, for example, the redshift factors or even the geometry. I mention this because, from the examples, it was not clear to me how to get the photon ring contributions and how data is stored when doing so,

(From: examples/mino-time-example)

`geometries = (Krang.ConeGeometry(θs, (i, rmin, rmax)) for i in 0:2)`

as opposed to

```
emission_coordinates!(coordinates, camera, τ)

time = coordinates[1,:,:]
radius = coordinates[2,:,:]
inclination = coordinates[3,:,:]
azimuth = mod2pi.(coordinates[4,:,:])
```
I understand the quantities are different here, but it would be beneficial to make them more explicit or even comment on them more.

- I could not completely understand the example "Raytracing a polygon mesh."  Could you please add a few sentences to introduce that example and comment it more? What do you mean by "Raytrace the mesh embedded in the Kerr space time with additive blending"?
