### 2024-10-09 499ee9c
Remove unused code in Kerr API
