+++ src/Krang.jl
@@ -21,7 +21,6 @@ include("cameras/SlowLightIntensityCamera.jl")
 include("cameras/IntensityCamera.jl")
 include("metrics/Kerr/misc.jl")
 include("metrics/Kerr/emission_coordinates.jl")
-include("metrics/Kerr/api.jl")
 include("materials/physicsUtils.jl")
 include("materials/ElectronSynchrotronPowerLawPolarization.jl")
 include("materials/ElectronSynchrotronPowerLawIntensity.jl")
