+++ src/cameras/IntensityCamera.jl
@@ -18,6 +18,25 @@ struct IntensityPixel{T} <: AbstractPixel{T}
     θo::T
     η::T
     λ::T
+    @doc """
+        IntensityPixel(met::Kerr{T}, α::T, β::T, θo::T) where {T}
+
+    Construct an `IntensityPixel` object with the given Kerr metric, screen coordinates, and inclination.
+
+    # Arguments
+    - `met::Kerr{T}`: The Kerr metric.
+    - `α::T`: The Bardeen α value (screen coordinate).
+    - `β::T`: The Bardeen β value (screen coordinate).
+    - `θo::T`: The inclination angle.
+
+    # Returns
+    - An `IntensityPixel` object initialized with the given parameters.
+
+    # Details
+    This function calculates the η and λ values using the provided Kerr metric and screen coordinates. 
+    It then computes the radial roots and adjusts them if necessary. 
+    Finally, it initializes an `IntensityPixel` object with the calculated values and the provided parameters.
+    """
     function IntensityPixel(met::Kerr{T}, α::T, β::T, θo::T) where {T}
         tempη = Krang.η(met, α, β, θo)
         tempλ = Krang.λ(met, α, θo)
@@ -40,7 +59,7 @@ end
 """
     $TYPEDEF
 
-Screen made of Intensity Pixels.
+Screen made of `IntensityPixel`s.
 """
 struct IntensityScreen{T, A <:AbstractMatrix} <: AbstractScreen
     "Minimum and Maximum Bardeen α values"
@@ -58,6 +77,26 @@ struct IntensityScreen{T, A <:AbstractMatrix} <: AbstractScreen
         β = βmin + (βmax - βmin) * (T(J)-1) / (res-1)
         screen[I, J] = IntensityPixel(met, α, β, θo)
     end
+
+    @doc """
+        IntensityScreen(met::Kerr{T}, αmin::T, αmax::T, βmin::T, βmax::T, θo::T, res::Int; A=Matrix) where {T}
+
+    Creates an intensity screen for the given Kerr metric. 
+    This camera caches information for fast light computations.
+
+    # Arguments
+    - `met::Kerr{T}`: The Kerr metric.
+    - `αmin::T`: Minimum value of α.
+    - `αmax::T`: Maximum value of α.
+    - `βmin::T`: Minimum value of β.
+    - `βmax::T`: Maximum value of β.
+    - `θo::T`: Observer's inclination angle. θo ∈ (0, π).
+    - `res::Int`: Resolution of the screen.
+    - `A=Matrix`: Optional argument to specify the type of matrix to use. A GPUMatrix can be used for GPU computations.
+
+    # Returns
+    - `IntensityScreen{T, A}`: An intensity screen object.
+    """
     function IntensityScreen(met::Kerr{T}, αmin::T, αmax::T, βmin::T, βmax::T, θo::T, res; A=Matrix) where {T}
         screen = A(Matrix{IntensityPixel{T}}(undef, res, res))
 
@@ -72,7 +111,7 @@ end
 """
     $TYPEDEF
 
-Observer sitting at radial infinity.
+Camera that caches fast light raytracing information for an observer sitting at radial infinity.
 The frame of this observer is alligned with the Boyer-Lindquist frame.
 """
 struct IntensityCamera{T, A} <: AbstractCamera
@@ -81,6 +120,25 @@ struct IntensityCamera{T, A} <: AbstractCamera
     screen::IntensityScreen{T, A}
     "Observer screen_coordinate"
     screen_coordinate::NTuple{2, T}
+
+    @doc """
[…]
