+++ src/metrics/Kerr/Kerr.jl
@@ -12,6 +12,18 @@ struct Kerr{T} <: AbstractMetric
     mass::T  
     "a = J/M, where J is the angular momentum and M is the mass of the black hole."
     spin::T
+
+    @doc """
+        Kerr(spin::T) where {T}
+    
+    Constructs a `Kerr` object representing the Kerr metric.
+    
+    # Arguments
+    - `spin::T`: The spin parameter `a = J/M`, where `J` is the angular momentum and `M` is the mass of the black hole. spin ∈ (-1, 0) ∪ (0, 1).
+    
+    # Returns
+    - A `Kerr` object with the given spin and a default mass of 1.
+    """
     function Kerr(spin::T) where {T}
         new{T}(one(T), spin)
     end

[…]
