### 2024-10-09 dominic-chang/Krang.jl#20 @dominic-chang
> Thank you both (@dominic-chang and @ExpandingMan) for the clarification. I think it is still worth clearly mentioning the dependencies all the examples need (e.g., CairoMakie or GLMakie), as that is indeed a standard practice.

I see your point, but I don't think this is standard practice. I think it is highly discouraged by the community to include heavy packages as dependencies in general. I believe this is the reason why 'Base' packages and 'PkgExtensions' exist in julia ecosystem. [Here's an example](https://discourse.julialang.org/t/dependency-policy-should-we-avoid-dependencies-or-embrace-them/8793/15) thread on the julia discourse where it is strongly advised to not include plotting packages as dependencies. Admittedly, this was a few years ago, but I think the guidelines are still relevant.

> - I indeed think you may want to comment a little bit more on the examples. Going back to the coordinate example, e.g., small things like just commenting on these variables will be helpful (which you are doing only in the second example):

Thanks for that suggestion. I have added some extra comments to the examples and additional doc strings. The docstrings of the constructors should mention the bounds for which the types are defined.

> - I think it would be very useful if you also provided an explicit example of how to export the results of the raytracing and other results. Users may want to export, for example, the redshift factors or even the geometry. I mention this because, from the examples, it was not clear to me how to get the photon ring contributions and how data is stored when doing so,

I'm sorry. I'm not sure I understand this point. Do you mean to include an example that saves ray-traced information to disk?

> - I could not completely understand the example "Raytracing a polygon mesh." Could you please add a few sentences to introduce that example and comment it more? What do you mean by "Raytrace the mesh embedded in the Kerr space time with additive blending"?

I've updated the example with text and added an animation that will hopefully better explain what's happening.

Changes have been made to the [dev docs](https://dominic-chang.com/Krang.jl/dev)
