### 2024-10-09 dominic-chang/Krang.jl#20 @alejandroc137
- I understand what you are sauing. I am not saying it should be installed when people get the package. Just that mention somewhere that the examples you are providing requiere CairoMakie and GLMakie. Without them, the examples won't work.

- Yes, I meant to include a simple example where the raytracing information can be extracted and used in other applications of your raytracing code.

I do not have more comments.
