### 2024-10-10 2c1d4c9
Update getting_started.md to mention plotting dependencies in examples
