### 2024-10-10 dominic-chang/Krang.jl#20 @dominic-chang
@alejandroc137  I'm still working on implementing all of your suggestions. Do you have any specific ideas for interfacing with `Krang` with external packages in mind? If not, I can just write an example that exports images to some possibly space efficient format.
