### 2024-10-10 dominic-chang/Krang.jl#20 @alejandroc137
I was not thinking about images. My comment was more about the idea that since Krang is a raytracing code, there should be a simple build-in routine to get/store the raytracing results. So, I was thinking of the following, which will be a minimal extension of what appears in the example coordinate_example, that instead of plotting the information, it saves the output.

So, for a given geometry, say a double cone, `ConeGeometry`, one can get the raytracing information for each pixel in the image plane. That is, for each pixel location (\alpha, \beta, in Bardeen coordinates), one gets the information of where the rays landed in that specific geometry, e.g.,:

\alpha, \beta -> r_n, t_n, theta_n, phi_n

Where the sub-label n relates to the photon ring index, With that information, people can then produce images, study geometries, describe a fluid on that geometry, get redshifts, etc. As I mentioned before, it was not very clear to me the best way to do it with the current code, as it seems to be different from using:

`Krang.ConeGeometry(θs, (i, rmin, rmax)) for i in 0:2)`

or

`emission_coordinates!(coordinates, camera, τ)`
