### 2024-10-10 dominic-chang/Krang.jl#20 @dominic-chang
I'm not really understanding what your intent was with the code examples you provided. If I understand what you mean, then your saying that I should, for example, save an array of coordinates to disk for the sub images generated from raytracing some geometry?

I think separately, I should also add an example of defining custom materials to be painted onto the geometries, and exporting that information as well. For example, a power law emission profile that returns pixel based intensities and redshifts which can then be exported to disk
