+++ dev/assets/examples_raytracing-mesh-example.md.iZRln6Mc.js
@@ -1,4 +1,4 @@
-import{_ as h,c as n,a2 as k,j as s,a,o as t}from"./chunks/framework.CCA2VXdB.js";const l="/Krang.jl/dev/assets/mesh_geometry_example.DdlC6w_K.png",p="/Krang.jl/dev/assets/mesh.CSDVnAi2.mp4",C=JSON.parse('{"title":"Raytracing a polygon mesh","description":"","frontmatter":{},"headers":[],"relativePath":"examples/raytracing-mesh-example.md","filePath":"examples/raytracing-mesh-example.md","lastUpdated":null}'),e={name:"examples/raytracing-mesh-example.md"};function E(r,i,g,d,y,F){return t(),n("div",null,i[0]||(i[0]=[k(`<h1 id="Raytracing-a-polygon-mesh" tabindex="-1">Raytracing a polygon mesh <a class="header-anchor" href="#Raytracing-a-polygon-mesh" aria-label="Permalink to &quot;Raytracing a polygon mesh {#Raytracing-a-polygon-mesh}&quot;">​</a></h1><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
+import{_ as h,c as n,a2 as k,j as s,a,o as t}from"./chunks/framework.CCA2VXdB.js";const l="/Krang.jl/dev/assets/mesh_geometry_example.CLWv3mW9.png",p="/Krang.jl/dev/assets/mesh.pMuoOyCG.mp4",C=JSON.parse('{"title":"Raytracing a polygon mesh","description":"","frontmatter":{},"headers":[],"relativePath":"examples/raytracing-mesh-example.md","filePath":"examples/raytracing-mesh-example.md","lastUpdated":null}'),e={name:"examples/raytracing-mesh-example.md"};function E(r,i,g,d,y,F){return t(),n("div",null,i[0]||(i[0]=[k(`<h1 id="Raytracing-a-polygon-mesh" tabindex="-1">Raytracing a polygon mesh <a class="header-anchor" href="#Raytracing-a-polygon-mesh" aria-label="Permalink to &quot;Raytracing a polygon mesh {#Raytracing-a-polygon-mesh}&quot;">​</a></h1><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
 <span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">import</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GLMakie </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">as</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GLMk</span></span>
 <span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeometryBasics</span></span>
[…]
