### 2024-10-10 dominic-chang/Krang.jl#20 @dominic-chang
> -  I think it is worth adding some notebooks instead of only scripts, at least for the most basic examples. This will allow for more comments and embedded figures that give a sense of what is going on without even running the code.

I added some[ notebook examples](https://github.com/dominic-chang/Krang.jl/tree/main/notebook_examples) of the scripts. Some of the media files are apparently too large to be rendered in the notebook preview on github, so I'm a bit worried about the extra overhead that is added to the git history. The information should match the [markdown](https://dominic-chang.com/Krang.jl/dev/examples/coordinate-example) that's generated in the docs.
