### 2024-10-10 #7273 @lrnv
@alejandroc137 Are your questions answered and requirements met by the author yet, or are you still waiting for responses / modifications / issues to be closed ? In the first case, maybe you could re-consider checking the boxes in this thread?

In both cases, you don't have much more to do, except maybe telling me briefly what did you think about the submission ?
