### 2024-10-11 43f98b5
Update Polarization material definitons. Removed UnionGeometry type
