+++ src/raytracer_api.jl
@@ -1,5 +1,0 @@
-export render
-
-function render!(::RayTrace, camera::AbstractCamera, scene::Scene)
-    mapreduce(mesh -> mesh.material.(camera.screen.pixels, Ref(mesh.geometry)), +, scene)
-end
