### 2024-10-11 184e274
Add new example that defines a custom materian and saves the output
