+++ docs/make.jl
@@ -48,7 +48,6 @@ makedocs(;
         ],
         "api.md",
     ],
-    checkdocs=:none
 )
 
 deploydocs(;
