+++ examples/Project.toml
@@ -12,6 +12,7 @@ JacobiElliptic = "2a8b799e-c098-4961-872a-356c768d184c"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 Krang = "54806c32-d51a-438d-8447-e0041be2fbfb"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
+NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationOptimisers = "42dfb2eb-d2b4-4451-abcd-913932933ac1"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
