+++ docs/Project.toml
@@ -14,6 +14,7 @@ Krang = "54806c32-d51a-438d-8447-e0041be2fbfb"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 MeshIO = "7269a6da-0436-5bbc-96c2-40638cbb6118"
+NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationOptimisers = "42dfb2eb-d2b4-4451-abcd-913932933ac1"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
