### 2024-10-11 844f3de
Update package references and acknowledgements
