### 2024-10-11 57bf646
Bump version to 0.3.0 in Project.toml
Update README.md

Add media for notebook examples

Fix typo in custom-material-example.jl

Update dependencies in Project.toml files

Update Julia setup actions in CI and Documentation workflows

Update notebook examples
