### 2024-10-11 f1de597
Bump version to 0.3.0 in Project.toml
