### 2024-10-11 #7273 @dominic-chang
@lrnv I've made changes to the draft. Do you what command I should use to update the draft with the new version?
