+++ dev/assets/examples_raytracing-mesh-example.md.BuC-Gh93.js
@@ -1,24 +1,4 @@
-<!DOCTYPE html>
-<html lang="en-US" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <title>Raytracing a polygon mesh | Krang.jl</title>
-    <meta name="description" content="Kerr Raytracer for Analytic Null Geodesics">
-    <meta name="generator" content="VitePress v1.4.0">
-    <link rel="preload stylesheet" href="/Krang.jl/dev/assets/style.DL6Cunng.css" as="style">
-    
-    <script type="module" src="/Krang.jl/dev/assets/app.Bbg4zgYJ.js"></script>
-    <link rel="preload" href="/Krang.jl/dev/assets/inter-roman-latin.Di8DUHzh.woff2" as="font" type="font/woff2" crossorigin="">
-    <link rel="modulepreload" href="/Krang.jl/dev/assets/chunks/theme.D3xGG602.js">
-    <link rel="modulepreload" href="/Krang.jl/dev/assets/chunks/framework._7uKzKn0.js">
-    <link rel="modulepreload" href="/Krang.jl/dev/assets/examples_raytracing-mesh-example.md.BuC-Gh93.lean.js">
-    <link rel="icon" href="/Krang_logo.ico">
-    <script id="check-dark-mode">(()=>{const e=localStorage.getItem("vitepress-theme-appearance")||"auto",a=window.matchMedia("(prefers-color-scheme: dark)").matches;(!e||e==="auto"?a:e==="dark")&&document.documentElement.classList.add("dark")})();</script>
-    <script id="check-mac-os">document.documentElement.classList.toggle("mac",/Mac|iPhone|iPod|iPad/i.test(navigator.platform));</script>
-  </head>
-  <body>
-    <div id="app"><div class="Layout" data-v-a9a9e638><!--[--><!--]--><!--[--><span tabindex="-1" data-v-c3508ec8></span><a href="#VPContent" class="VPSkipLink visually-hidden" data-v-c3508ec8> Skip to content </a><!--]--><!----><header class="VPNav" data-v-a9a9e638 data-v-f1e365da><div class="VPNavBar" data-v-f1e365da data-v-822684d1><div class="wrapper" data-v-822684d1><div class="container" data-v-822684d1><div class="title" data-v-822684d1><div class="VPNavBarTitle has-sidebar" data-v-822684d1 data-v-28a961f9><a class="title" href="/Krang.jl/dev/" data-v-28a961f9><!--[--><!--]--><!--[--><img class="VPImage logo" src="/Krang.jl/dev/Krang_logo.png" width="24" height="24" alt data-v-35a7d0b8><!--]--><span data-v-28a961f9>Krang.jl</span><!--[--><!--]--></a></div></div><div class="content" data-v-822684d1><div class="content-body" data-v-822684d1><!--[--><!--]--><div class="VPNavBarSearch search" data-v-822684d1><!--[--><!----><div id="local-search"><button type="button" class="DocSearch DocSearch-Button" aria-label="Search"><span class="DocSearch-Button-Container"><span class="vp-icon DocSearch-Search-Icon"></span><span class="DocSearch-Button-Placeholder">Search</span></span><span class="DocSearch-Button-Keys"><kbd class="DocSearch-Button-Key"></kbd><kbd class="DocSearch-Button-Key">K</kbd></span></button></div><!--]--></div><nav aria-labelledby="main-nav-aria-label" class="VPNavBarMenu menu" data-v-822684d1 data-v-e6d46098><span id="main-nav-aria-label" class="visually-hidden" data-v-e6d46098> Main Navigation </span><!--[--><!--[--><a class="VPLink link VPNavBar
[…]
