+++ dev/assets/examples_custom-material-example.md.DuNEBy_L.lean.js
@@ -1,0 +1,55 @@
+import{_ as l,c as n,a2 as a,j as i,a as t,o as h}from"./chunks/framework._7uKzKn0.js";const k="/Krang.jl/dev/assets/redshifts.Bo4g0lBq.png",C=JSON.parse('{"title":"Defining a Custom Material and Saving Output","description":"","frontmatter":{},"headers":[],"relativePath":"examples/custom-material-example.md","filePath":"examples/custom-material-example.md","lastUpdated":null}'),e={name:"examples/custom-material-example.md"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.023ex",height:"1.581ex",role:"img",focusable:"false",viewBox:"0 -677 1778 699","aria-hidden":"true"},d={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},E={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.186ex"},xmlns:"http://www.w3.org/2000/svg",width:"8.426ex",height:"1.808ex",role:"img",focusable:"false",viewBox:"0 -717 3724.1 799","aria-hidden":"true"};function g(y,s,o,c,F,m){return h(),n("div",null,[s[7]||(s[7]=a(`<h1 id="Defining-a-Custom-Material-and-Saving-Output" tabindex="-1">Defining a Custom Material and Saving Output <a class="header-anchor" href="#Defining-a-Custom-Material-and-Saving-Output" aria-label="Permalink to &quot;Defining a Custom Material and Saving Output {#Defining-a-Custom-Material-and-Saving-Output}&quot;">​</a></h1><p>We will define a redshift material for raytracing, and export the raytraced quantities to a .npy file.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span></code></pre></div><p>Materials should be functors (i.e. types with callable objects). Our material will raytrace the redshifts associated with a zero angular momentum observer (ZAMO) on a cone for a given sub-image. If the cone is self obscuring, then only the redshift on the side that is closest to the observer will be raytraced.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">struct</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> ZAMORedshifts{T} </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">&lt;:</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Krang.AbstractMaterial</span></span>
[…]

[…]
