### 2024-10-12 dominic-chang/Krang.jl#24 @ExpandingMan
difficulty with GPU examples
I'm still having quite a bit of trouble adapting the provided examples, in this case `coordinate-example.jl` to work with CUDA.

May I suggest the following: set up a complete working example that you have verified works properly, exactly as-is with whatever GPU you have access to, and then I will go and do whatever is needed to make sure it also works with CUDA?
