### 2024-10-12 #7273 @ExpandingMan
I'm pretty much ready to sign off on this and finish up my involvement, but I'd like to address the following:

First, while I have been able to verify that the bulk of the code here is appropriate for use on GPU, and can be made to work, at time of writing I'm still struggling to produce straightforward examples of common use cases.  I intend to work with the author (see [here](https://github.com/dominic-chang/Krang.jl/issues/24)) and hopefully with my access to nvidia GPU's I can get it the last bit of the way to working smoothly on CUDA, but I think a more blindingly obvious GPU example is needed first.

Second, what are the expectations here as far as rigorous validation of the integrals?  The math looks good to me, but there is of course a gaping chasm between reading the paper that derives the geodesic formulae with a cursory review of the code and a thorough and rigorous validation.  The author seems to have done quite a good job with testing, but even carefully reviewing the tests I think will take some effort.  The implementation requires quite a lot of subtlety as analytic solutions go because there is a bunch of branching and topology going on.  My understanding is that expectations regarding these kinds of detailed and rigorous validations vary from field to field and journal to journal, so I wanted to get a feel for what those expectations would be here.  (Btw, I suspect that maybe the authors of 1910.12881 have either informed or checked the work here?  They would likely be far more capable of spotting issues than any of us.)
