### 2024-10-12 dominic-chang/Krang.jl#24 @ExpandingMan
Specifically what I just did is try to run `coordinate-example.jl` from latest `main` by swapping out `rendered_scene` with the provided CUDA version.  Again, I really think a complete working as-is example would make things much easier.  Something even more minimal than `coordinate-example.jl` would be fine, I just want to be absolutely sure it's supposed to work on GPU in the first place before I try to get it working with CUDA.
