+++ v0.3.0/assets/examples_coordinate-example.md.CUBH94Oh.js
@@ -1,0 +1,72 @@
+import{_ as k}from"./chunks/coordinate.CXV4n4zL.js";import{_ as l,c as n,a2 as a,j as i,a as h,o as t}from"./chunks/framework.Bb15l1tD.js";const u=JSON.parse('{"title":"Raytracing with inclination","description":"","frontmatter":{},"headers":[],"relativePath":"examples/coordinate-example.md","filePath":"examples/coordinate-example.md","lastUpdated":null}'),p={name:"examples/coordinate-example.md"},e={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.355ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.999ex",height:"1.95ex",role:"img",focusable:"false",viewBox:"0 -705 883.6 862.1","aria-hidden":"true"};function E(d,s,g,y,o,F){return t(),n("div",null,[s[4]||(s[4]=a(`<h1 id="Raytracing-with-inclination" tabindex="-1">Raytracing with inclination <a class="header-anchor" href="#Raytracing-with-inclination" aria-label="Permalink to &quot;Raytracing with inclination {#Raytracing-with-inclination}&quot;">​</a></h1><p>This example shows how to access coordinate information from the raytracing process. You will likely need to do this when making custom physics <code>materials</code>. We will raytrace a sequence of cones in the region around a Kerr black hole as seen by an observer stationed at infinity. We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons that are emitted from the source, at a fixed inclination angles with respect to the black hole&#39;s spin axis.</p><p>First, let&#39;s import Krang and CairoMakie for plotting.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Krang</span></span>
+<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
+<span class="line"></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">curr_theme </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Theme</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    Axis </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> (</span></span>
[…]
