### 2024-10-12 aa0fdd0
Refactor emission_coordinates function to fix type instability
