### 2024-10-12 dominic-chang/Krang.jl#24 @dominic-chang
Ah, yeah I see. There's a typo in the docs, it should be

`Array(coordinate_point.(CuArray(camera.screen.pixels), Ref(geometry))`

not just

`coordinate_point.(CuArray(camera.screen.pixels), Ref(geometry)`

since I have to take it off the GPU to plot.

Were you having an indexing error? Here's a version of the `draw` function that worked with Metal on my mac

```
# Draw Function
# This function draws the coordinates associated with the n=0,1,2 subimages of a cone with opening angle θs.
function draw!(axes_list, camera, coordinates, rmin, rmax, θs)
    times, radii, azimuths = coordinates
    map(axes -> empty!.(axes), axes_list)

    geometries = (Krang.ConeGeometry(θs, (i, rmin, rmax)) for i in 0:2)

    for (i, geometry) in enumerate(geometries)
        rendered_scene = Array(coordinate_point.(MtlArray(camera.screen.pixels), Ref(geometry)))
        for I in CartesianIndices(rendered_scene)
            times[I] = rendered_scene[I][1]
            radii[I] = rendered_scene[I][2]
            azimuths[I] = rendered_scene[I][4]
        end
        coordinates = (times, radii, mod2pi.(azimuths ))
        for j in 1:3
            heatmap!(axes_list[i][j], coordinates[j], colormap = colormaps[j], colorrange=colorrange[j])
        end
    end
end
```
I think replacing the `MtlArray` with `CuArray` should do the trick. I would also try the main branch. There was a small typo that was causing a type instability whose fix I just pushed.
