### 2024-10-12 dominic-chang/Krang.jl#24 @ExpandingMan
Could you please, please add a full and explicit example, even a very minimal one, that definitely works with GPU exactly as is?  I don't even care if it's in the docs, and I will gladly do whatever is required to make sure it works for me swapping out whatever you use with CUDA, but I've done quite a lot of juggling these examples trying to adapt them and I just want to firmly ground in a common frame of reference.
