### 2024-10-12 #7273 @dominic-chang
On point 2:
The development of this code happened independently from the authors of arXiv:1910.12881 unfortunately. I'm not even sure if they know this code exists. The code also implements some additional integrals that are outside the scope of the original paper. One thing I will say is that there are tests in the CI that check the analytic solutions by comparing them to numerical implementations of the integrals from the paper. See for example: [ @testset "Radial Integrals 1" begin](https://github.com/dominic-chang/Krang.jl/blob/aa0fdd05440875ac209087c22d014e32840c62ad/test/kerr_misc_tests.jl#L18). This way I can somewhat guarantee the solutions are correct, assuming that I copied down the integral expressions correctly.
