### 2024-10-12 #7273 @ExpandingMan
All perfectly fair.  I think you did everything you could have done here, it's just a matter of how much additional work are we all going to do to validate it.  I kind of expect that with JOSS a sanity check is enough, and digging through the details to the point where I have as good a chance spotting a mistake as you do will take quite a lot of work, so I'm satisfied.  I just wanted to make clear first that no, I have not stripped down and reconstructed the integral code to the point where I can personally vouch for its accuracy and second, to make sure we all have the same expectations.  Again, short of doing a lot of work, everything looks good to me.
