### 2024-10-12 #7273 @alejandroc137
The author has implemented all my comments and suggestions. This code and the applications are a great addition to our community. I echo one @ExpandingMan comment that it was hard to get easy examples of common use cases from the ones the author provided. However, the new examples are more commented on and have more parts that could serve as the building blocks for more applications.

I think the equations and implementation of the raytracing must be, for the most part, correct (a careful checking of them will be very time-consuming indeed). Suppose there are (significant) typos or these are incorrect. In that case, in my experience, they would already be manifest in the results: with these analytical expressions (which we also implemented in [AART](https://github.com/iAART/aart) with my collaborators, including one of the authors of 1910.12881), when mainly considering ordinary geodesics (i.e., η >0), if there were extra/missing terms, sign errors, lots of nan will appear, as these functions are very sensible, and the geometry of the black hole extreme enough to saturate their bounds, and higher order rings will look "weird" and these errors will be apparent. The current implementation's novel part is how the time was normalized, and the results make physical sense in the regions I checked. At any rate, if there were any typos, the code can be easily fixed, and I am confident the author will certainly do so.

My last super minor comment to the manuscript is to change "... yet to be observed" to "...yet to be resolved" when talking about photon rings. As the author surely knows, the photons from these higher-order contributions do arrive at the radio telescopes, but the whole point of these types of efforts and future instruments is to understand them enough to resolve them!
