### 2024-10-12 #7273 @lrnv
Yes @ExpandingMan you are right, the level of validation I can see here seem largely enough for this kind of review. After all, reviewers are not authors, and bugs are part of the life cycle of software. I would not be able to comment on the details though as I don't fully understand what is going on here, but you three seem to know what you are talking about.

@ExpandingMan what is the status of https://github.com/dominic-chang/Krang.jl/issues/24 then ? Do you want to wait for resolution before acceptign the paper ?

@alejandroc137 Well received, thanks a lot for your input on this submission.

@dominic-chang, let us also wait to hear @aplavin oppinon
