### 2024-10-12 #7273 @ExpandingMan
Great.  In that case I would like to get a GPU example fully squared away and then I'll be all done.
