+++ gpu_examples/Project.toml
@@ -1,0 +1,5 @@
+[deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+Krang = "54806c32-d51a-438d-8447-e0041be2fbfb"
+Metal = "dde4c033-4e86-420c-a63e-0dd931031962"
