### 2024-10-12 dominic-chang/Krang.jl#24 @dominic-chang
Sorry, I'm really not sure why I wasn't understanding that you were asking me to add examples. I think I must be sleep deprived 😅

Here's some example scripts that run on my GPU. I think they should run if you just replace the `MtlArray` with `CuArray`.
Please use the main branch for now. I'll push a new patch in a bit.

https://github.com/dominic-chang/Krang.jl/tree/main/gpu_examples
