### 2024-10-12 b38ec6a
Refactor paper.md to correct typo in photon-rings description
