### 2024-10-12 #7273 @dominic-chang
> My last super minor comment to the manuscript is to change "... yet to be observed" to "...yet to be resolved" when talking about photon rings. As the author surely knows, the photons from these higher-order contributions do arrive at the radio telescopes, but the whole point of these types of efforts and future instruments is to understand them enough to resolve them!

Great! This change has been made
