### 2024-10-12 9117e0c
Add generate_rays method that allows GPU execution on ray generation.
Update mino-time-example-metal to use generate_rays
