+++ examples/mino-time-example.jl
@@ -1,32 +1,12 @@
 # # Rays parameterized with Mino time
 
 # In this example, we will raytrace the region around a Kerr black hole as seen by an observer stationed at infinity.
-# We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons as they are emitted from the 
-# source, at a fixed inclination angle from the black hole's spin axis.
+# We will return the coordinates associated with a ray by marching along the ray's Mino time parameter from the assymptotic observer.
 #
 # First, let's import Krang and CairoMakie for plotting.
 using Krang
 import GLMakie as GLMk
 GLMk.Makie.inline!(true)
-
-#
-# We will use a 0.99 spin Kerr black hole viewed by an asymptotic observer at an inclination angle of θo=π/4. 
-# A region spanned by radii between the horizon and 20M at varying inclinations will be raytraced onto the 20Mx20M 
-# screen of the observer.
-metric = Krang.Kerr(0.99); # Kerr spacetime with 0.99 spin
-θo = 85 * π / 180; # Observer inclination angle with respect to spin axis
-sze = 200; # Number of pixels along each axis of the screen
-rmin = Krang.horizon(metric); # Minimal radius to raytrace
-rmax = 10; # Maximal radius to raytrace
-ρmax = 5; # Size of the screen
-
-# We will define a camera with the above parameters.
-# The SlowLightIntensityCamera pre-calculates information about the spacetime and the observer's screen to speed up the 
-# raytracing for slowlight applications.
-
-camera = Krang.SlowLightIntensityCamera(metric, θo, -ρmax, ρmax, -ρmax, ρmax, sze);
-
-# Let us now create a figure to plot the emission coordinates on,
 
 curr_theme = GLMk.Theme(# Makie theme
     fontsize=20,
@@ -45,9 +25,27 @@ curr_theme = GLMk.Theme(# Makie theme
 
 GLMk.set_theme!(GLMk.merge(curr_theme, GLMk.theme_latexfonts()))
 
+# 
+# We will use a 0.99 spin Kerr black hole viewed by an asymptotic observer at an inclination angle of θo=π/4. 
+# 
+
+metric = Krang.Kerr(0.99); # Kerr spacetime with 0.99 spin
+θo = 85 * π / 180; # Observer inclination angle with respect to spin axis
+sze = 200; # Number of pixels along each axis of the screen
+ρmax = 5; # Size of the screen
+
+# We will define a camera with the above parameters.
+# The SlowLightIntensityCamera pre-calculates information about the spacetime and the observer's screen to speed up the 
+# raytracing for slow light applications.
+
+camera = Krang.SlowLightIntensityCamera(metric, θo, -ρmax, ρmax, -ρmax, ρmax, sze);
+
+# Let us now create a figure to plot the emission coordinates on.
+
 fig = GLMk.Figure(resolution=(500, 600));
-# and use this figure make an animation by looping over the inclination angle θs.
-# This loop will plot the emission coordinates for each θs.
+
+# We will create a loop to plot the emission coordinates for each θs.
+
 recording = GLMk.record(fig, "raytrace.gif", range(0.1, 3, length=290), framerate=15) do τ
     GLMk.empty!(fig)
 
@@ -61,7 +59,7 @@ recording = GLMk.record(fig, "raytrace.gif", range(0.1, 3, length=290), framerat
     data = (time, radius, inclination, azimuth)
[…]
