+++ paper/paper.bib
@@ -1,109 +1,0 @@
-@misc{Bezanson2015,
-      title={Julia: A Fresh Approach to Numerical Computing}, 
-      author={Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
-      year={2015},
-      eprint={1411.1607},
-      archivePrefix={arXiv},
-      primaryClass={cs.MS},
-      doi = {10.1137/141000671}
-}
-@article{JohnsonRing,
-author = {Michael D. Johnson  and Alexandru Lupsasca  and Andrew Strominger  and George N. Wong  and Shahar Hadar  and Daniel Kapec  and Ramesh Narayan  and Andrew Chael  and Charles F. Gammie  and Peter Galison  and Daniel C. M. Palumbo  and Sheperd S. Doeleman  and Lindy Blackburn  and Maciek Wielgus  and Dominic W. Pesce  and Joseph R. Farah  and James M. Moran },
-title = {Universal interferometric signatures of a black hole\&\#x2019;s photon ring},
-journal = {Science Advances},
-volume = {6},
-number = {12},
-pages = {eaaz1310},
-year = {2020},
-doi = {10.1126/sciadv.aaz1310},
-URL = {https://www.science.org/doi/abs/10.1126/sciadv.aaz1310},
-eprint = {https://www.science.org/doi/pdf/10.1126/sciadv.aaz1310},
-abstract = {A black hole image contains a series of increasingly narrow rings measurable with progressively longer interferometers. The Event Horizon Telescope image of the supermassive black hole in the galaxy M87 is dominated by a bright, unresolved ring. General relativity predicts that embedded within this image lies a thin “photon ring,” which is composed of an infinite sequence of self-similar subrings that are indexed by the number of photon orbits around the black hole. The subrings approach the edge of the black hole “shadow,” becoming exponentially narrower but weaker with increasing orbit number, with seemingly negligible contributions from high-order subrings. Here, we show that these subrings produce strong and universal signatures on long interferometric baselines. These signatures offer the possibility of precise measurements of black hole mass and spin, as well as tests of general relativity, using only a sparse interferometric array.}}
-@article{M87PaperI,
-doi = {10.3847/2041-8213/ab0ec7},
-url = {https://dx.doi.org/10.3847/2041-8213/ab0ec7},
-year = {2019},
-month = {apr},
-publisher = {The American Astronomical Society},
-volume = {875},
-number = {1},
-pages = {L1},
-author = {{Event Horizon Telescope Collaboration}},
-title = {First M87 Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole},
-journal = {The Astrophysical Journal Letters},
[…]
