+++ examples/coordinate-example.jl
@@ -25,7 +25,7 @@ curr_theme = Theme(
 set_theme!(merge!(curr_theme, theme_latexfonts()))
 
 # We will use a 0.99 spin Kerr black hole viewed by an asymptotic observer at an inclination angle of θo=π/4. 
-# A region spanned by radii between the horizon and 20M at varying inclinations will be raytraced onto the 20Mx20M 
+# A region spanned by radii between the horizon and 10M at varying inclinations will be raytraced onto the 20Mx20M 
 # screen of the observer.
 
 
@@ -108,5 +108,5 @@ end
 # ```julia
 # using CUDA
 # 
-# rendered_scene = coordinate_point.(CuArray(camera.screen.pixels), Ref(geometry))
+# rendered_scene = Array(coordinate_point.(CuArray(camera.screen.pixels), Ref(geometry)))
 # ```
