+++ gpu_examples/README.MD
@@ -1,3 +1,5 @@
 # GPU examples
 
 Here are a collection of GPU versions of the examples from the example script folder.
+> :warning:
+The underlying algorithms are designed for `Float64`, so there may be numerical issues present when using `Float32`.

[…]
