### 2024-10-13 1411153
Fix bug with total mino time calculation in generate ray
Refactor IntensityPixel and SlowLightIntensityPixel to store total mino time of ray upon definition
