### 2024-10-13 0adc2b3
Fix Latex Rendering bug
