### 2024-10-13 6b02014
Bump version to 0.3.1 in Project.toml
