+++ dev/assets/chunks/@localSearchIndexroot.ml_iwP5a.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":46,"nextId":46,"documentIds":{"0":"/Krang.jl/dev/api#Krang-api","1":"/Krang.jl/dev/api#Raytracing-Functions","2":"/Krang.jl/dev/api#Metric-Functions","3":"/Krang.jl/dev/api#Radial-Integrals","4":"/Krang.jl/dev/api#Angular-Integrals","5":"/Krang.jl/dev/api#Screen-Coordinates","6":"/Krang.jl/dev/api#Local-Frame-transformations","7":"/Krang.jl/dev/api#polarization","8":"/Krang.jl/dev/api#Meshes,-Geometries-and-Materials","9":"/Krang.jl/dev/api#Pixel-Related-Functions","10":"/Krang.jl/dev/api#misc","11":"/Krang.jl/dev/examples/coordinate-example#Raytracing-with-inclination","12":"/Krang.jl/dev/examples/custom-material-example#Defining-a-Custom-Material-and-Saving-Output","13":"/Krang.jl/dev/examples/mino-time-example#Rays-parameterized-with-Mino-time","14":"/Krang.jl/dev/examples/neural-net-example#Neural-Network-Emission-Model-Example","15":"/Krang.jl/dev/examples/neural-net-example#Fitting-the-model","16":"/Krang.jl/dev/examples/polarization-example#Creating-a-Custom-Dual-Cone-Model","17":"/Krang.jl/dev/examples/raytracing-mesh-example#Raytracing-a-polygon-mesh","18":"/Krang.jl/dev/getting_started#What-is-Krang.jl?","19":"/Krang.jl/dev/getting_started#philosophy","20":"/Krang.jl/dev/getting_started#meshes","21":"/Krang.jl/dev/getting_started#materials","22":"/Krang.jl/dev/getting_started#geometries","23":"/Krang.jl/dev/getting_started#raytracing","24":"/Krang.jl/dev/getting_started#Raytracing-conical-surfaces","25":"/Krang.jl/dev/getting_started#Raytracing-with-rays-parameterized-by-Mino-time","26":"/Krang.jl/dev/getting_started#cameras","27":"/Krang.jl/dev/getting_started#examples","28":"/Krang.jl/dev/kerr_geodesic_summary#The-Geodesics-of-the-Kerr-Metric","29":"/Krang.jl/dev/kerr_geodesic_summary#Geodesic-Equations","30":"/Krang.jl/dev/kerr_geodesic_summary#Hamilton-Jacobi-Equations","31":"/Krang.jl/dev/kerr_geodesic_summary#Solutions-by-Quadrature","32":"/Krang.jl/dev/kerr_geodesic_summary#references","33":"/Krang.jl/dev/polarization#linear-polarization","34":"/Krang.jl/dev/polarization#Linear-Polarization-of-null-geodesics-in-the-Newmann-Penrose-Formalism","35":"/Krang.jl/dev/polarization#Walker-Penrose-Constant-and-Killing-Spinors-of-type-D-spacetimes","36":"/Krang.jl/dev/polarization#references","37":"/Krang.jl/dev/newmann_penrose#Newmann-Penrose-Formalism","38":"/Krang.jl/dev/newmann_p
[…]

[…]
