### 2024-10-13 dominic-chang/Krang.jl#24 @dominic-chang
I pushed a new patch. The GPU examples aren't in the documentation itself, but they are available as scripts in the /gpu_examples folder as previously mentioned. Let me know if there are any issues. There are sometimes rendering artifacts with `Float32` so I would use `Float64` if you can. The rendering issues, I'm pretty sure, are due to incompatible convergence criteria in some of the routines I'm using. Here's an example where the rays develop spikes at some locations.
