+++ v0.3.1/assets/chunks/coordinate.DPZoYlbx.js
@@ -1,0 +1,1 @@
+const s="/Krang.jl/v0.3.1/assets/coordinate.BHWT5GLC.gif";export{s as _};
