### 2024-10-13 dominic-chang/Krang.jl#24 @ExpandingMan
So, unfortunately the GPU compiler seems to very easily go type unstable for a lot of this stuff, even though it's all perfectly stable on CPU.  It looks like all the tuple return types are making it shit itself.  Based on my experience, *maybe* things would be better if there were more dedicated structs for holding a lot of the return values? But who knows.

Regardless, it's now looking to me like, although some stuff works, it would take someone with an nvidia card a significant effort to go through and get it working with CUDA.  That's unfortunate but it's not the fault of this package.  Definitely would not make sense for that to hold up the JOSS review, in my mind that's done and I'll go to the issue and check my list and sign off now.

As for the floating point stuff, if there's any low-hanging fruit of stuff you can make more robust in obvious ways, by all means go for it, but floating point is just one of those facts of life that we all have to deal with and you will not be able to rescue users from its fundamental limitations, so I think this is fine.  If there's any algorithms that you think are particularly dangerous or safe with `Float32`, might be nice to point it out in the docs, but it's certainly not your job to explain floating point to users so I wouldn't be worried about it.
