### 2024-10-13 #7273 @ExpandingMan
Looks good to me, I have checked off all boxes and am signing off.  Thanks for this contribution Dominic!

(On a side note, I have discovered that getting this running with CUDA is a lot harder than I had hoped.  The GPU compiler makes a lot of these functions type unstable for some reason even though it's all fine on GPU.  *This is not the fault of this package*, but it does mean it would take someone with access to an nvidia card significant effort to get this package working smoothly with CUDA.  Would have been nice to see those examples working on CUDA, but I don't think this has any bearing on this review, and there's not really anything different Dominic could have done here, GPU's are just a big pain in the ass.)
