### 2024-10-13 dominic-chang/Krang.jl#24 @dominic-chang
Thanks for that. I'm surprised you're having so many issues with CUDA. Did the GPU examples fail when you tried them, or are you having issue with other functions?
