### 2024-10-13 dominic-chang/Krang.jl#24 @ExpandingMan
Naively it fails with inscrutable errors, but as I dug into it I realized that it was type unstable and that was causing the failures.  Again, I see this as "GPU's are a pain in the ass", there are undoubtedly things that can be done to get it working, but it's not because this package is doing anything bad or unusual, it's just that CUDA is too fragile.

It's looking to me like most of the instability is happening around the `Tuple` returns from many of the functions in this package.  Why those are so fragile I do not know.

Most of the code in this package looks to me entirely appropriate for GPU (no heap allocs, type stable on CPU at least) so I don't think there are any glaring fundamental issues, it would just take someone to go through the pain of getting it to work.  It's doable, so I don't think it's a big deal that CUDA doesn't work in the current state.
