### 2024-10-15 6f8c10a
Restore joss paper, bib, and workflow yml
