### 2024-10-15 005a387
fix typos in JOSS Paper
