### 2024-10-15 #7273 @dominic-chang
> - Double check authors and affiliations (including ORCIDs)

Checked and Confirmed

> - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Done. The version for the paper is v0.3.1

> - Archive the release on Zenodo/figshare/etc and post the DOI here.

Done. Here is the doi 10.5281/zenodo.13936258

> - Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Checked

> - Make sure that the license listed for the archive is the same as the software license.

Both Licenses have been set to the MIT license
