### 2024-10-15 #7273 @dominic-chang
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07273/joss.07273/10.21105.joss.07273.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.07273/joss.07273/10.21105.joss.07273.pdf) 📄 👈

Corrected some typos. Here's the new version
