### 2024-10-17 #7273 @lrnv
@dominic-chang there seem to be a missing DOI, and the automatic check found a potential one, could you check if it is indeed the right doi (and then add it in the bib file) or if it's a false positive ? Then I'll recomand-accept and pass the hand to the TEiC for finalisation
