### 2024-10-17 21a7e38
Update BHEX citation. in paper.bib
