+++ paper/paper.bib
@@ -62,14 +62,20 @@ journal = {The Astrophysical Journal Letters},
 abstract = {We present the first Event Horizon Telescope (EHT) observations of Sagittarius A* (Sgr A*), the Galactic center source associated with a supermassive black hole. These observations were conducted in 2017 using a global interferometric array of eight telescopes operating at a wavelength of λ = 1.3 mm. The EHT data resolve a compact emission region with intrahour variability. A variety of imaging and modeling analyses all support an image that is dominated by a bright, thick ring with a diameter of 51.8 ± 2.3 μas (68% credible interval). The ring has modest azimuthal brightness asymmetry and a comparatively dim interior. Using a large suite of numerical simulations, we demonstrate that the EHT images of Sgr A* are consistent with the expected appearance of a Kerr black hole with mass ∼4 × 106  M ⊙, which is inferred to exist at this location based on previous infrared observations of individual stellar orbits, as well as maser proper-motion studies. Our model comparisons disfavor scenarios where the black hole is viewed at high inclination (i &gt; 50°), as well as nonspinning black holes and those with retrograde accretion disks. Our results provide direct evidence for the presence of a supermassive black hole at the center of the Milky Way, and for the first time we connect the predictions from dynamical measurements of stellar orbits on scales of 103–105 gravitational radii to event-horizon-scale images and variability. Furthermore, a comparison with the EHT results for the supermassive black hole M87* shows consistency with the predictions of general relativity spanning over three orders of magnitude in central mass.}
 }
 @misc{lupsasca2024blackholeexplorerphoton,
-      title={The Black Hole Explorer: Photon Ring Science, Detection and Shape Measurement}, 
-      author={Alexandru Lupsasca and Alejandro Cárdenas-Avendaño and Daniel C. M. Palumbo and Michael D. Johnson and Samuel E. Gralla and Daniel P. Marrone and Peter Galison and Paul Tiede and Lennox Keeble},
-      year={2024},
-      eprint={2406.09498},
-      archivePrefix={arXiv},
-      primaryClass={gr-qc},
-      url={https://arxiv.org/abs/2406.09498}, 
+author = {Alexandru Lupsasca and Alejandro C{\'a}rdenas-Avenda{\~n}o and Daniel C. M. Palumbo and Michael D. Johnson and Samuel E. Gralla and Daniel P. Marrone and Peter Galison and Paul Tiede and Lennox Keeble},
+title = {{The Black Hole Explorer: photon ring science, detection, and shape measurement}},
+volume = {13092},
+booktitle = {Space Telescopes and Instrumentation 2024: Optical, Infrared, and Millimeter Wave},
+editor = {Laura E. Coyle and Shuji Matsuura and Marshall D. Perrin},
+organization = {International Society for Optics and Photonics},
+publisher = {SPIE},
+pages = {130926Q},
+keywords = {black holes, photon ring, interferometry, VLBI, space telescope, AGN, EHT},
+year = {2024},
+doi = {10.1117/12.3019437},
+URL = {https://doi.org/10.1117/12.3019437}
 }
+
 @article{cardenas-avendano,
   title = {Adaptive analytical ray tracing of black hole photon rings},
   author = {C\'ardenas-Avenda\~no, Alejandro and Lupsasca, Alexandru and Zhu, Hengrui},
