### 2024-10-17 #7273 @dominic-chang
Great, thanks. I updated the bib file with the new doi in commit [44912ade1672ed7ec585c51c907423a0944f3667](https://github.com/dominic-chang/Krang.jl/commit/44912ade1672ed7ec585c51c907423a0944f3667)
