### 2024-10-17 #7273 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
