### 2024-10-17 7545652
Update paper text and acknowledgments
