### 2024-10-18 #7273 @lrnv
@dominic-chang Nothing ! We simply have to wait for the TEic to take the lead now, should take a few days max (depends on his/her bandwidth, which i do not control).

Please sit tight and wait a bit more :)
