### 2024-10-21 #7273 @dominic-chang
Not sure if it's an issue or not, but this pull request is closed. Should it be reopened?

https://github.com/openjournals/joss-papers/pull/6016
