### 2024-10-23 #7273 @warrickball
Hi @dominic-chang, I'm now getting around to this but there'll be a few more days of details before we're fully published. A few comments to start:

* In the [Zenodo archive](https://zenodo.org/records/13936258), the right sidebar says "Version v1", rather than v0.3.1, which is what's tagged here and indicated in the actual zip file and tarball. Could you fix that v1?
* Could add the country to the affiliations? "USA" will do.
* You've added a footnote for the [Julia homepage](https://julialang.org). JOSS style is to rather [use inline hyperlinks](https://joss.readthedocs.io/en/latest/paper.html#links):

> Open Journal publications are not limited by the constraints of print publications. We encourage authors to use hyperlinks for websites and other external resources.

I'll take a closer look at the reference list in due course but note these two jumped out at me:

* If I [follow the DOI](https://epubs.siam.org/doi/10.1137/141000671) for the *Bezanson et al. (2015)* reference for Julia, I can get a BibTeX entry that includes journal, volume and page number information, and is also a *2017* publication, not 2015. I'm not sure if you're using a different version of this entry?
* The title of *Event Horizon Telescope Collaboration (2022)* has lost all capitalization. We use sentence case for reference titles in JOSS but you may need to protect e.g. `{Sagittarius} {A}* {Event Horizon Telescope}` and `{Milky Way}`.
* *Johnson et al. (2020)* has some garbled character, presumably meant to be some form of apostrophe.

If you happen to be able to fix those before I have a close look at the source, that'd be great, but I'll almost certainly end up opening a PR with other editorial fixes in the next few days.
