### 2024-10-23 b1d0e75
Fix paper affiliations and citations
