+++ paper/paper.bib
@@ -1,12 +1,17 @@
-@misc{Bezanson2015,
-      title={Julia: A Fresh Approach to Numerical Computing}, 
-      author={Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
-      year={2015},
-      eprint={1411.1607},
-      archivePrefix={arXiv},
-      primaryClass={cs.MS},
-      doi = {10.1137/141000671}
+@article{Bezanson2017,
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+URL = { https://doi.org/10.1137/141000671},
+eprint = { https://doi.org/10.1137/141000671},
+abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
 }
+
 @article{JohnsonRing,
 author = {Michael D. Johnson  and Alexandru Lupsasca  and Andrew Strominger  and George N. Wong  and Shahar Hadar  and Daniel Kapec  and Ramesh Narayan  and Andrew Chael  and Charles F. Gammie  and Peter Galison  and Daniel C. M. Palumbo  and Sheperd S. Doeleman  and Lindy Blackburn  and Maciek Wielgus  and Dominic W. Pesce  and Joseph R. Farah  and James M. Moran },
 title = {Universal interferometric signatures of a black hole\&\#x2019;s photon ring},
@@ -57,7 +62,7 @@ volume = {930},
 number = {2},
 pages = {L12},
 author = {{Event Horizon Telescope Collaboration}},
-title = {First Sagittarius A* Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole in the Center of the Milky Way},
+title = {First {Sagittarius} {A}* {Event Horizon Telescope} Results. I. The Shadow of the Supermassive Black Hole in the Center of the {Milky Way}},
 journal = {The Astrophysical Journal Letters},
[…]
