### 2024-10-23 4af85da
Update paper.bib titles for consistency
