+++ paper/paper.bib
@@ -34,7 +34,7 @@ volume = {875},
 number = {1},
 pages = {L1},
 author = {{Event Horizon Telescope Collaboration}},
-title = {First M87 Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole},
+title = {First M87 {Event Horizon Telescope} Results. I. The Shadow of the Supermassive Black Hole},
 journal = {The Astrophysical Journal Letters},
 abstract = {When surrounded by a transparent emission region, black holes are expected to reveal a dark shadow caused by gravitational light bending and photon capture at the event horizon. To image and study this phenomenon, we have assembled the Event Horizon Telescope, a global very long baseline interferometry array observing at a wavelength of 1.3 mm. This allows us to reconstruct event-horizon-scale images of the supermassive black hole candidate in the center of the giant elliptical galaxy M87. We have resolved the central compact radio source as an asymmetric bright emission ring with a diameter of 42 ± 3 μas, which is circular and encompasses a central depression in brightness with a flux ratio ≳10:1. The emission ring is recovered using different calibration and imaging schemes, with its diameter and width remaining stable over four different observations carried out in different days. Overall, the observed image is consistent with expectations for the shadow of a Kerr black hole as predicted by general relativity. The asymmetry in brightness in the ring can be explained in terms of relativistic beaming of the emission from a plasma rotating close to the speed of light around a black hole. We compare our images to an extensive library of ray-traced general-relativistic magnetohydrodynamic simulations of black holes and derive a central mass of M = (6.5 ± 0.7) × 109 M⊙. Our radio-wave observations thus provide powerful evidence for the presence of supermassive black holes in centers of galaxies and as the central engines of active galactic nuclei. They also present a new tool to explore gravity in its most extreme limit and on a mass scale that was so far not accessible.}
 }
@@ -48,7 +48,7 @@ volume = {875},
 number = {1},
 pages = {L2},
 author = {{Event Horizon Telescope Collaboration}},
-title = {First M87 Event Horizon Telescope Results. II. Array and Instrumentation},
+title = {First M87 {Event Horizon Telescope} Results. II. Array and Instrumentation},
 journal = {The Astrophysical Journal Letters},
[…]
