### 2024-10-23 34db8e6
Update paper.bib title for consistency
