+++ paper/paper.bib
@@ -14,7 +14,7 @@ abstract = { Bridging cultures that have often been distant, Julia combines expe
 
 @article{JohnsonRing,
 author = {Michael D. Johnson  and Alexandru Lupsasca  and Andrew Strominger  and George N. Wong  and Shahar Hadar  and Daniel Kapec  and Ramesh Narayan  and Andrew Chael  and Charles F. Gammie  and Peter Galison  and Daniel C. M. Palumbo  and Sheperd S. Doeleman  and Lindy Blackburn  and Maciek Wielgus  and Dominic W. Pesce  and Joseph R. Farah  and James M. Moran },
-title = {Universal interferometric signatures of a black hole\&\#x2019;s photon ring},
+title = {Universal interferometric signatures of a black hole's photon ring},
 journal = {Science Advances},
 volume = {6},
 number = {12},
