### 2024-10-23 9dfca17
Update paper.bib title for consistency
