+++ paper/paper.bib
@@ -105,7 +105,7 @@ volume = {939},
 number = {2},
 pages = {107},
 author = {Daniel C. M. Palumbo and Zachary Gelles and Paul Tiede and Dominic O. Chang and Dominic W. Pesce and Andrew Chael and Michael D. Johnson},
-title = {{Bayesian Accretion Modeling}: Axisymmetric Equatorial Emission in the Kerr Spacetime},
+title = {{Bayesian Accretion Modeling}: Axisymmetric Equatorial Emission in the {Kerr} Spacetime},
 journal = {The Astrophysical Journal},
 abstract = {The Event Horizon Telescope (EHT) has produced images of two supermassive black holes, Messier 87* (M 87*) and Sagittarius A* (Sgr A*). The EHT collaboration used these images to indirectly constrain black hole parameters by calibrating measurements of the sky-plane emission morphology to images of general relativistic magnetohydrodynamic (GRMHD) simulations. Here, we develop a model for directly constraining the black hole mass, spin, and inclination through signatures of lensing, redshift, and frame dragging, while simultaneously marginalizing over the unknown accretion and emission properties. By assuming optically thin, axisymmetric, equatorial emission near the black hole, our model gains orders of magnitude in speed over similar approaches that require radiative transfer. Using 2017 EHT M 87* baseline coverage, we use fits of the model to itself to show that the data are insufficient to demonstrate existence of the photon ring. We then survey time-averaged GRMHD simulations fitting EHT-like data, and find that our model is best-suited to fitting magnetically arrested disks, which are the favored class of simulations for both M 87* and Sgr A*. For these simulations, the best-fit model parameters are within ∼10% of the true mass and within ∼10° for inclination. With 2017 EHT coverage and 1% fractional uncertainty on amplitudes, spin is unconstrained. Accurate inference of spin axis position angle depends strongly on spin and electron temperature. Our results show the promise of directly constraining black hole spacetimes with interferometric data, but they also show that nearly identical images permit large differences in black hole properties, highlighting degeneracies between the plasma properties, spacetime, and, most crucially, the unknown emission geometry when studying lensed accretion flow images at a single frequency.}
 }
