### 2024-10-29 #7273 @warrickball
I've proposed some editorial fixes in [a PR](https://github.com/dominic-chang/Krang.jl/pull/25). Note I haven't suggested a correction to this sentence, which needs fixing (emphasis mine):
> This feature could potentially be seen in the near future **with a recent as a space extension** to the EHT aimed of detection and measurement (Lupsasca et al., 2024).

It isn't unambiguous to me what the intended meaning is so I'll leave it to you to fix. Let me know when you've done so and I'll do another dry run to check the metadata that's generated for publication.
