+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - black hole 
   - general relativity
   - astronomy
-  - raytracing
+  - ray-tracing
 authors:
   - name: Dominic Chang
     orcid: 0000-0001-9939-5257
@@ -21,7 +21,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-`Krang` is a [Julia](https://julialang.org) [@Bezanson2017] package that implements efficient algorithms for raytracing emission geometries in the Kerr black hole space time.
+`Krang` is a [Julia](https://julialang.org) [@Bezanson2017] package that implements efficient algorithms for ray tracing emission geometries in the Kerr black hole space time.
 It is GPU compatible and is specialized for studies of sub-image contributions from gravitational lensed sources [@JohnsonRing]. 
 Such algorithms are of interest for modeling the sources seen by Very Long Baseline Interferometry (VLBI) observations of Low Luminosity Active Galactic Nuclei (LLAGN) such as those imaged by the Event Horizon Telescope Collaboration (EHTC).
 
@@ -35,21 +35,21 @@ The large scale of the black holes allows for many of these effects to be descri
 A relativistic image feature that has been theorized to exist, but is yet to be resolved, are the individual sub-image contributions to the overall image structure known as photon-rings, [@JohnsonRing].
 Photon rings are of particular interest because of their strong dependence on gravitational effects and their insensitivity to variations in the emission physics around the black hole.
 The observation of a photon ring would therefore serve as effective probe for measurements of black hole characteristics like spin or help facilitate tests of gravity.
-This feature could potentially be seen in the near future with a recent as a space extension to the EHT aimed of detection and measurement [@lupsasca2024blackholeexplorerphoton].
+This feature could potentially be seen in the near future with a recently proposed space extension to the EHT aimed at detection and measurement [@lupsasca2024blackholeexplorerphoton].
 
 Scientific studies of black hole images within the current software landscape often requires great compromise due to the computational complexity of the problem.
-Other difficultes can arise from interfacing existing raytracing software with the wider ecosystem of tools; it is difficult, for example, to apply machine learning optimization algorithms to existing python raytracing codes since they typically rely on special functions that are not implemented within current machine learning frameworks.
+Other difficultes can arise from interfacing existing ray tracing software with the wider ecosystem of tools; it is difficult, for example, to apply machine learning optimization algorithms to existing python ray tracing codes since they typically rely on special functions that are not implemented within current machine learning frameworks.
 Existing python implementations are also bounded to CPU evaluations, limiting their capability of accessing acceleration from specialized hardware.
-A Julia implementation of raytracing algorithms is thus beneficial because of the languages modular design, differentiable programming and efficient execution.
+A Julia implementation of ray tracing algorithms is thus beneficial because of the languages modular design, differentiable programming and efficient execution.
 Krang therefore benefits from the 'plug and play' nature of the julia programming language, allowing for easy development and synergy with existing analysis pipelines, and fast CPU/GPU executions.
 
 
 # Similar Packages
 - `AART` [@cardenas-avendano]: An adaptive analytical ray tracing code for geodesics in the Kerr space time in python.
 
-- `KerrBam` [@Palumbo_2022]: An analytical raytracing code for equatorial synchrotron models in python.
+- `KerrBam` [@Palumbo_2022]: An analytical ray tracing code for equatorial synchrotron models in python.
 
-- `Gradus` [@Gradus]: A julia implementation of a numeric, general relativistic raytracer with radiative transfer.
+- `Gradus` [@Gradus]: A julia implementation of a numeric, general relativistic ray tracer with radiative transfer.
 
 # Acknowledgements
[…]
