### 2024-10-30 5c0c6ba
Update paper title for consistency
