+++ src/materials/ElectronSynchrotronPowerLawPolarization.jl
@@ -77,7 +77,7 @@ end
     - `R::T`: The characteristic radius of the emissivity profile.
     - `p1::T`: The first power-law index.
     - `p2::T`: The second power-law index.
-    - `subimgs::NTuple{N, Int}`: The sub-images to raytrace.
+    - `subimgs::NTuple{N, Int}`: The sub-images to ray trace.
 """
 struct ElectronSynchrotronPowerLawPolarization{N, T} <: AbstractMaterial 
     magnetic_field::SVector{3, T}
@@ -116,7 +116,7 @@ struct ElectronSynchrotronPowerLawPolarization{N, T} <: AbstractMaterial
     - `R::T`: The characteristic radius of emissivity profile.
     - `p1::T`: The first power-law index.
     - `p2::T`: The second power-law index.
-    - `subimgs::NTuple{N, Int}`: The sub-images to raytrace.
+    - `subimgs::NTuple{N, Int}`: The sub-images to ray trace.
 
     # Returns
     - `ElectronSynchrotronPowerLawPolarization{N, T}`: A new instance of `ElectronSynchrotronPowerLawPolarization`.
