+++ examples/coordinate-example.jl
@@ -2,7 +2,7 @@
 
 # This example shows how to access coordinate information from the raytracing process.
 # You will likely need to do this when making custom physics `materials`.
-# We will raytrace a sequence of cones in the region around a Kerr black hole as seen by an observer stationed at infinity.
+# We will ray trace a sequence of cones in the region around a Kerr black hole as seen by an observer stationed at infinity.
 # We will show the emission coordinates of the n=0 (direct) and n=1 (indirect) photons that are emitted from the 
 # source, at a fixed inclination angles with respect to the black hole's spin axis.
 #
@@ -25,15 +25,15 @@ curr_theme = Theme(
 set_theme!(merge!(curr_theme, theme_latexfonts()))
 
 # We will use a 0.99 spin Kerr black hole viewed by an asymptotic observer at an inclination angle of θo=π/4. 
-# A region spanned by radii between the horizon and 10M at varying inclinations will be raytraced onto the 15x15 
+# A region spanned by radii between the horizon and 10M at varying inclinations will be ray traced onto the 15x15 
 # screen of the observer.
 
 
 metric = Krang.Kerr(0.99); # Kerr metric with a spin of 0.99
 θo = 45 * π / 180; # inclination angle of the observer. θo ∈ (0, π)
 sze = 400; # resolution of the screen is sze x sze
-rmin = Krang.horizon(metric); # minimum radius to be raytraced
-rmax = 10.0; # maximum radius to be raytraced
+rmin = Krang.horizon(metric); # minimum radius to be ray traced
+rmax = 10.0; # maximum radius to be ray traced
 ρmax = 15.0; # horizontal and vertical limits of the screen
 
 # Create Figure
