+++ dev/assets/index.md.BUSt_jlW.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,c as e,o as a}from"./chunks/framework.aGcT7TsX.js";const p=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"KRANG.jl","text":"Kerr Ray-tracer for Analytic Null Geodesics","tagline":"A performant accurate ray tracer for problems in the Kerr spacetime.","image":{"src":"/Krang_logo.png","alt":"Krang.jl"},"actions":[{"theme":"brand","text":"Getting Started","link":"/getting_started"},{"theme":"alt","text":"View on Github","link":"https://github.com/dominic-chang/Krang.jl"},{"theme":"alt","text":"API","link":"/api"}]},"features":[{"icon":"<img width=\\"64\\" height=\\"64\\" src=\\"https://metal.juliagpu.org/stable/assets/logo.png\\" />","title":"GPU Compatible","details":"Type stable and type preserving. GPU compatible with CUDA.jl and Metal.jl."},{"icon":"<img width=\\"64\\" height=\\"64\\" src=\\"https://enzyme.mit.edu/julia/stable/assets/logo.svg\\" alt=\\"markdown\\"/>","title":"Automatic Differentiable, Backwards Propagation compatible","details":"Supports automatic differentiation with Enzyme.jl"}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),i={name:"index.md"};function r(n,o,l,s,c,m){return a(),e("div")}const g=t(i,[["render",r]]);export{p as __pageData,g as default};

[…]
