### 2024-10-31 #7273 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
