### 2024-10-31 #7273 @warrickball
Thanks once again to @aplavin, @alejandroc137 & @ExpandingMan for reviewing and @lrnv for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.

@dominic-chang, your paper has been published in JOSS! :tada:
